% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbott.R
\docType{class}
\name{abbott-class}
\alias{abbott-class}
\alias{plot,abbott-method}
\alias{print,abbott-method}
\alias{show,abbott-method}
\title{A S4 class object for the Abbott-Firestone data}
\usage{
\S4method{print}{abbott}(x)

\S4method{show}{abbott}(object)

\S4method{plot}{abbott}(x, y, ..., ynew = NULL)
}
\arguments{
\item{x}{An \code{abbott} object.}

\item{object}{An \code{abbott} object.}

\item{y}{unnused argument of \code{plot} generic function.}

\item{...}{Optional arguments to be passed to \code{plot} function}

\item{ynew}{Optional probability points where to calculate heights.}
}
\description{
A S4 class object for the Abbott-Firestone data
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: Custom print for \code{abbott} objects

\item \code{show}: Custom show for \code{abbott} objects

\item \code{plot}: Custom plot for \code{abbott} objects
}}
\section{Slots}{

\describe{
\item{\code{dist}}{distribuition function for the Abbott object}

\item{\code{Sa}}{Arithmetic average of absolute values}

\item{\code{Sq}}{Root mean squared}

\item{\code{Sv}}{Maximum valley depth}

\item{\code{Sp}}{Maximum peak height}

\item{\code{St}}{Maximum Height of the surface}

\item{\code{Ssk}}{Skewness}

\item{\code{Sku}}{kurtosis}
}}

