% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbott.R
\name{abbott}
\alias{abbott}
\title{Abbott-Firestone data constructor}
\usage{
abbott(surf)
}
\arguments{
\item{surf}{a \code{cimg} object}
}
\value{
A S4 object of class \code{\link{abbott-class}}
}
\description{
This function creates an \code{abbott} object based on a surface
}
\examples{
file <- system.file("extdata", "ground.txt", package = "surf")
surf <- read.surf(file)
surf <- level(surf)
abb <- abbott(surf)
print(abb)
plot(abb)
}

