% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{abs.max}
\alias{abs.max}
\title{L-infinity norm}
\usage{
\method{abs}{max}(x, ...)
}
\arguments{
\item{x}{a \code{numeric} vector.}

\item{...}{other parameters for \link{which.max}.}
}
\value{
maximum in absolute value
}
\description{
L-infinity norm
}
