% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{addGeneTx}
\alias{addGeneTx}
\title{Add identifiers of genes and transcripts
Fill up missing "gene" and "transcript" in genome annotation
If the genome annotation lacks "gene" and "transcript" entries, add to it.}
\usage{
addGeneTx(anno)
}
\arguments{
\item{anno}{a \code{GRanges} object of genome annotation, from \link{import}.}
}
\value{
a \code{GRanges} object with added \code{gene_id} and \code{transcript_id} columns.
}
\description{
Add identifiers of genes and transcripts
Fill up missing "gene" and "transcript" in genome annotation
If the genome annotation lacks "gene" and "transcript" entries, add to it.
}
