% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{align}
\alias{align}
\title{Title}
\usage{
align(surf, targ)
}
\arguments{
\item{surf}{\code{cimg} object}

\item{targ}{\code{cimg} object}
}
\description{
description
}
\examples{
file <- system.file("extdata", "sur1.txt", package = "surf")
sur1 <- read.surf(file)
file <- system.file("extdata", "sur2.txt", package = "surf")
sur2 <- read.surf(file)
sur3 <- align(sur2, sur1)
par(mfrow = c(2,2), mar = c(2,2,1,1))
plot(sur1, asp = 1)
plot(sur2, asp = 1)
plot(sur3, asp = 1)
plot(error(sur1, sur3), asp = 1)
mean(error(sur1, sur3))/diff(range(sur1))
}

