% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{annotateEvent}
\alias{annotateEvent}
\title{annotate AS events (generate feature bins)}
\usage{
annotateEvent(
  isoPLas,
  cores = max(1, detectCores() - 2),
  min.event.length = 6,
  anno_ss = NULL,
  remove.duplicate = F
)
}
\arguments{
\item{isoPLas}{isoform parts list with AS events}

\item{cores}{# working cores}

\item{min.event.length}{numeric (positive), minimum length of a valid event}

\item{anno_ss}{GRanges of start/stop codon, UTR, Selenocysteine, etc; used to exclude some interior events (i.e., SE/RI/A3SS/A5SS)}

\item{remove.duplicate}{remove identical event duplicates (keep one randomly)}
}
\value{
a `surfParsing` object
}
\description{
annotate AS events (generate feature bins)
}
