% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{as.observations}
\alias{as.observations}
\title{transforms \code{imlist} objects into observation dataframes}
\usage{
as.observations(surf)
}
\arguments{
\item{surf}{\code{imlist} object}
}
\value{
\code{data.frame} of observations
}
\description{
Each \code{cimg} object inside \code{surf} is vectorized and alocated in an
individual line of a resultind data.frame
}
