% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{bruteCombn}
\alias{bruteCombn}
\title{brute-force choose, this is just a wraper of combn() function in utils}
\usage{
bruteCombn(set, ...)
}
\description{
brute-force choose, this is just a wraper of combn() function in utils
}
