% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{c_granges}
\alias{c_granges}
\title{Concatenate a list of GRanges}
\usage{
c_granges(grl, use.names = T, sep = ".", save.names = F)
}
\arguments{
\item{grl}{a list of \link{GRanges}.}

\item{use.names}{logical, whether (\code{TRUE}) to inherit list names.}

\item{sep}{character, separator between list names and GRanges names.}

\item{save.names}{logical or character, if not FALSE, save list names as an attribute (i.e., \code{\link[=mcols]{mcols()}}).s}
}
\value{
a \link{GRanges} object
}
\description{
Concatenate a list of GRanges
}
