% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/click_graph.R
\name{click_graph}
\alias{click_graph}
\title{Use shiny to return x- and y-coordinates when clicking on a plot}
\source{
https://shiny.posit.co/r/articles/build/plot-interaction/
}
\usage{
click_graph(x)
}
\arguments{
\item{x}{A saved plot}
}
\description{
Use shiny to return x- and y-coordinates when clicking on a plot
}
\examples{
#click_graph(plot(rnorm(10),rnorm(10)))

}
\author{
Chang W, Cheng J, Allaire J, Sievert C, Schloerke B, Xie Y, Allen J, McPherson J, Dipert A, Borges B (2022). \emph{shiny: Web Application Framework for R}. R package version 1.7.4, \url{https://CRAN.R-project.org/package=shiny}.
}
