% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{countLogicRle}
\alias{countLogicRle}
\title{Count number of `TRUE` from a logical rle}
\usage{
countLogicRle(x)
}
\arguments{
\item{x}{logical rle}
}
\value{
a integer rle
}
\description{
Count number of `TRUE` from a logical rle
}
