% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R, R/daseq.R
\docType{class}
\name{daseqResults-class}
\alias{daseqResults-class}
\alias{getAUC}
\alias{getAUC,daseqResults-method}
\title{DASeq results}
\usage{
getAUC(object)

\S4method{getAUC}{daseqResults}(object)
}
\description{
DASeq results
}
