% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{daseqResults}
\alias{daseqResults}
\title{DASeq Results}
\usage{
daseqResults(object)
}
\arguments{
\item{object}{a \code{surf} object output by \link{daseq}.}
}
\value{
a \code{daseqResults} object.
}
\description{
DASeq Results
}
