% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired-data.R
\docType{data}
\name{df1}
\alias{df1}
\alias{df2}
\alias{artificial}
\title{Artificial datasets}
\format{
Two objects of class \code{data.frame}: \code{df1, df2}.
}
\usage{
data("artificial")
}
\description{
Artificial paired datasets, obtained using a SRSWOR from a pseudo-population of 100 000 individuals.
}
\details{
The population was generated by this Markov chain:

\tabular{lllll}{
       \tab      \tab           \tab t1       \tab          \cr
       \tab      \tab  1        \tab 2        \tab 3        \cr
       \tab 1    \tab  .720     \tab .135     \tab .045     \cr
t0     \tab 2    \tab  .015     \tab .030     \tab .005     \cr
       \tab 3    \tab  .005     \tab .005     \tab .040     \cr
}

The \code{df1} contains the inclusion probabilities in the \code{prob} column.
}
\seealso{
\code{\link{pnadc1}} for real data from a Brazilian survey.
}
\keyword{artificial}
