% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired-data.R
\docType{data}
\name{dfa0}
\alias{dfa0}
\alias{dfa1}
\alias{artificial}
\title{Artificial datasets}
\format{Two objects of class \code{data.frame}: \code{dfa0, dfa1}.}
\usage{
data("artificial")
}
\description{
Artificial paired datasets, obtained using a SRSWOR from a pseudo-population of 1 000 000 individuals.
}
\details{
The population was generated by this Markov Chain:

\tabular{lllll}{
       \tab      \tab           \tab t1       \tab          \cr
       \tab      \tab  1        \tab 2        \tab 3        \cr
       \tab 1    \tab  .720     \tab .135     \tab .045     \cr
t0     \tab 2    \tab  .015     \tab .030     \tab .005     \cr
       \tab 3    \tab  .005     \tab .005     \tab .040     \cr
}

The \code{dfa0} contains the inclusion probabilities in the \code{prob} column.
}
\seealso{
\code{\link{dfr0}} for real data from a Brazilian survey.
}
\keyword{artificial}
\keyword{datasets}
