% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired-data.R
\docType{data}
\name{dfr0}
\alias{dfr0}
\alias{dfr1}
\alias{real}
\title{Real Datasets}
\format{Two objects of class \code{data.frame}: \code{dfr0, dfr0}}
\usage{
data("real")
}
\description{
Amazonas state subset of the Brazilian Continous Household Survey,
containing data from paired observations in the 1st and 2nd quarters of 2018.
}
\seealso{
\code{\link{dfa0}} for artificial data from a pseudo-population.
}
\keyword{real}
