% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daseq.R
\name{diffAUC}
\alias{diffAUC}
\title{Differential activity (via AUC)}
\usage{
diffAUC(
  rankings,
  targetSet,
  controlSet = NULL,
  n.sample = 1000,
  verbose = F,
  ...
)
}
\arguments{
\item{rankings}{'Rankings' created for this dataset with
\code{\link[AUCell]{AUCell_buildRankings}}.}

\item{verbose}{Should the function show progress messages? (TRUE / FALSE)}

\item{...}{Other arguments}

\item{size}{# of times of controlSet sampling.}
}
\description{
Detect differential activity using the AUC measure and RNA-seq quantification. This is a inference unit/engine of daseq.
This unit uses one worker only; daseq cordinates these units parallelly.
}
