% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drseq.R
\name{drseq}
\alias{drseq}
\title{DrSeq}
\usage{
drseq(drd, cores = max(1, detectCores() - 2), verbose = F, ...)
}
\arguments{
\item{drd}{a DEXSeqDataSet object created by `drseqData`.}

\item{cores}{integer, number of computing workers.}

\item{verbose}{logical, whether (`TRUE`) to echo progess.}

\item{...}{addtional parameters for `DEXSeq`.}
}
\value{
a `drseqResults` object.
}
\description{
Fit DrSeq models using the DEXSeq engine.
}
