% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drseq.R
\name{drseqData}
\alias{drseqData}
\title{DrSeq Data}
\usage{
drseqData(
  event,
  sampleData,
  design = ~sample + exon + condition:exon,
  remove.overlap.exon = F,
  anno.prefix = "drseq.annotation",
  anno.format = "gff2",
  minMQS = 10,
  isPairedEnd = T,
  cores = max(1, detectCores() - 2),
  verbose = F,
  ...
)
}
\arguments{
\item{event}{a `surf` object from [parseEvent].}

\item{sampleData}{data.frame, describes the RNA-seq samples, contains at least two columns -- `bam` and `condition`, the row.names represent sample names.}

\item{design}{
    A formula which specifies the design of the experiment. It must
    specify an interaction term between a variable from the sampleData
    columns with the 'exon' variable. By default, the design 
    will be '~ sample + exon + condition:exon'. This formula indicates
    the contrast between 'condition' and exon', i.e. differences in exon 
    usage due to changes in the 'condition' variable.  See the vignette 
    for more examples of other designs.
  }

\item{remove.overlap.exon}{logical, remove overlaping exons across genes (default to FALSE).}

\item{anno.prefix}{file names for outputting annotation files.}

\item{anno.format}{character, e.g. "gtf", as accepted by [rtracklayer::export].}

\item{cores}{integer, number of available workers, sent to `nthreads` of `featureCounts`.}

\item{verbose}{logical, whether (`TRUE`) to echo progess.}

\item{...}{additional parameters for [Rsubread::featureCounts].}

\item{minMQS;isPairedEnd}{as defined in [Rsubread::featureCounts]. Note that the default is customized for SURF (see details for more information).}
}
\value{
a `surf` object, with `drseqData` slot updated.
}
\description{
Create the `drseqData` slot needed by DrSeq analysis, which is a `DEXSeqDataSet`.
It is possible to output event read counts for later usage.
}
\details{
If you used Illumina HiSeq 2000, set `strandSpecific = 2` (reversed strand).
}
