% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired-data.R
\docType{data}
\name{f.qtr}
\alias{f.qtr}
\title{Brazilian Continous Household Survey Paired Data: 1st Quarter}
\format{An object of class \code{data.frame}.}
\usage{
f.qtr
}
\description{
Subset of the Brazilian Continous Household Survey for the Amazonas state,
containing data from paired observations in the 1st quarter.
}
\seealso{
\code{\link{s.qtr}}
}
\keyword{datasets}
