% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associate.R
\name{fa.plot}
\alias{fa.plot}
\title{Functional association plot}
\usage{
fa.plot(sres, plot.event = "all", trim = sres@params$trim, fdr.cutoff = 0.05)
}
\arguments{
\item{sres}{a `surfResults` object.}

\item{event}{a character vector, event type wanted. In particular, "all" means all event types.}
}
\value{
a ggplot object.
}
\description{
This function provides a ggplot implementation for functional association (FA Plot).
}
