% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faseq.R
\name{faseqFit}
\alias{faseqFit}
\title{Functional Association using Sequencing data}
\usage{
faseqFit(event, min.size = 60, trim = 0.025, verbose = F)
}
\arguments{
\item{event}{a \code{faseqData} obejcet.}

\item{min.size}{\code{integer}, the minimum size of "reliable" training set, default to `60`.}

\item{trim}{\code{numeric}, the percentile used to trim the training data. This is useful in producing a more robust estimation of functional association.}

\item{verbose}{\code{logical}, whether to print out progress information.}
}
\value{
a \code{surf} object with \code{faseqResults} slot updated.
}
\description{
This function performs functiona association test (FAT). 
The null hypothesis of FAT is that there is no association between feature signals and differential ATR.
}
