% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{faseqInfer}
\alias{faseqInfer}
\title{Functional association inference}
\usage{
faseqInfer(event, fdr.cutoff = 0.05, signal.cutoff = 20)
}
\arguments{
\item{event}{a \code{surf} object from \link{faseq} or \link{faseqFit}.}

\item{fdr.cutoff}{\code{numeric}, significance cutoff for the adjusted p-values of FAT.}

\item{signal.cutoff}{\code{numeric}, threshold cut-off for the eCLIP signals, default to 20. Set this to 0 if don't want to filter those location with low eCLIP signals of the RBP.}
}
\value{
a \code{surf} object, with one added \code{inferredFeature} column (inclusion/exclusion/none).
}
\description{
Inference the functionality of individual location features,
where the RBP is likely to interact and regulate the corresponding ATR event.
A location feature is inferred as function associated if
(i) it is included in FAT, and
(ii) the corresponding FAT is significant (padj < cut.off), and
(iii) it has strong binding signal (featureSignal > cut.off).
}
