% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{faseqResults}
\alias{faseqResults}
\title{FASeq Results}
\usage{
faseqResults(object)
}
\arguments{
\item{object}{a \code{surf} object output by \link{faseq} or \link{faseqFit} or \link{faseqInference}.}
}
\value{
a \code{faseqResults} object
}
\description{
FASeq Results
}
