% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associate.R
\name{fat}
\alias{fat}
\title{Functional Association Testing (FAT)}
\usage{
fat(event, min.size = 60, trim = 0.025, verbose = F, ...)
}
\arguments{
\item{event}{a `surfData` obejcet.}

\item{min.size}{integer, the minimum size of "reliable" training set, default to `60`.}

\item{trim}{numeric, trimmed quantile, for head and tail `trim` portion of sample will be ignored.}
}
\value{
a `surfResults` object
}
\description{
This function tests the functional association between feature signals and differential ATR.
}
