% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{featureCounts}
\alias{featureCounts}
\title{Customized \code{featureCounts}}
\usage{
featureCounts(..., verbose = F)
}
\arguments{
\item{...}{parameters for \code{featureCounts}.}

\item{verbose}{\code{logical}, whether to print out the progress information.}
}
\value{
See \link[Rsubread:featureCounts]{Rsubread::featureCounts} documentation.
}
\description{
This is a wrapper of \link[Rsubread:featureCounts]{Rsubread::featureCounts} allowing verbose suppression.
This redirects the output to \verb{/dev/null}, so it assumes a UNIX-like
system.
}
