% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_replace.R
\name{find_replace}
\alias{find_replace}
\title{Find and replace}
\usage{
find_replace(original, to.match, to.replace)
}
\arguments{
\item{original}{The vector to search}

\item{to.match}{A vector that possibly contain matches to original. Must be same length as to.replace}

\item{to.replace}{A vector of values to use when a match is found between original and to.match. Must be same length as to.match}
}
\value{
Returns a data frame showing the original vector and the replacement. If no match was found, the original be returned. Both columns are characters.
}
\description{
Performs a find replace on a vector by comparing to another vector. When no matches occur, the original value is retained.
}
\examples{
#E.g., Search "original" for "1", "2", "3", "4", and replace with "blue","green","purple","yellow"
original<-c(3,10,3,1,5,2)
to.find<-c(1,2,3,4)
to.replace<-c("blue","green","purple","yellow")
find_replace(original,to.find,to.replace)
}
