% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{findAdjacentFeature}
\alias{findAdjacentFeature}
\title{Find adjacent features of given feature}
\usage{
findAdjacentFeature(
  plas,
  body,
  type = c("lead", "lag"),
  depth.exon = 100,
  depth.intron = 300
)
}
\arguments{
\item{plas}{one isoform element in \code{isoPLas}.}

\item{body}{\code{list}, $id_tx, $event_id = body bins indices.}

\item{type}{\code{character}, either"lead" for upstream, or "lag" for downstream.}

\item{depth.exon}{\code{integer}, extended depth into exon, default 100 nt.}

\item{depth.intron}{\code{integer}, extended depth into intron, default 300 nt.}
}
\value{
a \code{GRangesList}.
}
\description{
Locate the adjacent bins at the desired direction and extract the genomic
bins, given some location features of AS events. This is a helper function to
\code{getFeature()}.
}
