% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{findBodyFeature}
\alias{findBodyFeature}
\title{Find body features}
\usage{
findBodyFeature(plas, type = c("lead", "lag"), depth.exon = 100)
}
\arguments{
\item{plas}{one element from isoPLas.}

\item{type}{character, either "lead" for upstream, or "lag" for downstream}

\item{depth.exon}{depth extended into exon, default is 100 nt}
}
\value{
a `GRangesList`
}
\description{
Locate the body feature of AS events and extract the genomic bins. 
This is a helper function to `getFeature()`
}
