% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{findNeighborFeature}
\alias{findNeighborFeature}
\title{Find features on the (upstream or downstream) constitutive exons}
\usage{
findNeighborFeature(
  plas,
  body,
  type = c("lead", "lag"),
  depth.exon = 100,
  constitutive = T
)
}
\arguments{
\item{plas}{one isoform element from isoPLas}

\item{body}{list, $id_tx$event_id = body bins indices}

\item{type}{character, either "lead" for upstream, or "lag" for downstream}

\item{depth.exon}{extended depth into exon}

\item{constitutive}{logical, whether to use consecutive or adjacent neighbor exon. The default is `TRUE`.}
}
\value{
a `GRangesList`
}
\description{
Locate the feature on the (upstream or downstream) constitutive exons and extract the genomic bins, given some body features of AS events. This is a helper function to `getFeature()`
}
