% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form.R
\name{form}
\alias{form}
\alias{form.cimg}
\alias{form.imlist}
\title{Form removal}
\usage{
form(surf, form = "paraboloid")

\method{form}{cimg}(surf, form = "paraboloid")

\method{form}{imlist}(surf, form = "paraboloid")
}
\arguments{
\item{surf}{a \code{\link[imager]{cimg}} or a \code{\link[imager]{imlist}}
object.}

\item{form}{an optional parameter that overides the information given on
formula. Options are: "ellipsoid", "paraboloid". #enable partial match}
}
\value{
a corrected \code{\link[imager]{cimg}} or a
  \code{\link[imager]{imlist}} object.
}
\description{
Removes the underlying format of a surface.
}
\examples{
surf <- ground[[1]]
par(mfrow = c(2,3))
plot(surf)
plot(form(surf))
plot(surf - form(surf))

plot(surf)
plot(form(surf, value ~ x + y))
plot(surf - form(surf, value ~ x + y))

}
