% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_df_names.R
\name{format_df_names}
\alias{format_df_names}
\title{Format data frame names}
\usage{
format_df_names(
  x,
  accents = TRUE,
  spaces = TRUE,
  lowercase = TRUE,
  excess.periods = TRUE
)
}
\arguments{
\item{x}{Data frame}

\item{accents}{Should accents be removed? Default = TRUE}

\item{spaces}{Should spaces be removed? Default = TRUE}

\item{lowercase}{Should all letters all be lowercase? Default = TRUE}

\item{excess.periods}{Should excess periods be removed? Default = TRUE}
}
\description{
Remove letter accents; replace spaces with periods; make all names lowercase; remove periods if they separate single characters
}
\examples{
fake<-as.data.frame(matrix(rep(1:4,7), ncol=7))
names(fake)<-c('    SOme DATA  here ','Données','n e b','A.S.O','.cfvn','ac.kg.tow','gh..kg.tow')
fake
format_df_names(fake)
format_df_names(fake,accents=FALSE)
format_df_names(fake,spaces=FALSE)
format_df_names(fake,lowercase=FALSE)
format_df_names(fake,excess.periods=FALSE)
}
