% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{genePartsList}
\alias{genePartsList}
\title{FASeq results}
\usage{
genePartsList(object)
}
\arguments{
\item{object}{a \code{surf} object output by \link{faseq} or \link{faseqFit} or \link{faseqInference}.}
}
\value{
a \code{DataFrame} object.
}
\description{
FASeq results
}
