% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daseq.R
\name{getControlSet}
\alias{getControlSet}
\title{Get control sets}
\usage{
getControlSet(event, targetSets, id_column = "transcript_id", verbose = F)
}
\value{
list of character, find a control set for each one of targetSets.
}
\description{
Get control sets
}
