% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{getEES}
\alias{getEES}
\title{Get exon end site from exon parts list}
\usage{
getEES(x)
}
\arguments{
\item{x}{an element of isoform part list.}
}
\description{
This is a helper function of \link{getEvent}.
}
