% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{getEvent}
\alias{getEvent}
\title{Find AS events from isoform parts list}
\usage{
getEvent(isoPL, cores = max(1, detectCores() - 2))
}
\arguments{
\item{isoPL}{isoform parts list}
}
\description{
Find AS events from isoform parts list
}
