% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{getFeature}
\alias{getFeature}
\title{Annotate AS events (generate feature bins)}
\usage{
getFeature(
  isoPLas,
  anno_event,
  depth.exon = 100,
  depth.intron = 300,
  cores = max(1, detectCores() - 2),
  remove.duplicate = T,
  verbose = T
)
}
\arguments{
\item{isoPLas}{isoform parts list with AS events}

\item{anno_event}{event annotation w/o features}

\item{depth.exon}{extended depth into exon, default 50 nt}

\item{depth.intron}{extended depth into intron, default 300 nt}

\item{cores}{# working cores}

\item{remove.duplicate}{remove identical event duplicates (keep one)}

\item{verbose}{logical, whether (`TRUE`) to echo progess}
}
\value{
a `surfParsing` object containing the `feature` attribute.
}
\description{
Annotated RBP features of AS event in `GRangesList` called `feature`
}
