% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associate.R
\name{getFeatureSignal}
\alias{getFeatureSignal}
\title{Obtain AS/ATS/APA event feature signal}
\usage{
getFeatureSignal(
  event,
  sampleData,
  signal.type = "FPKM",
  FUN.aggregate = "mean",
  output.signal = F,
  output.prefix = "feature",
  minMQS = 10,
  minOverlap = 12,
  isPairedEnd = T,
  cores = max(1, detectCores() - 2),
  verbose = F,
  ...
)
}
\arguments{
\item{event}{a `drseqResults` or `surfData` object, annotation of events}

\item{sampleData}{data.frame, must contain two columns -- `bam` and `condition` (for "IP" and "input", the IP should come first, for the ease of constrasting), row.names represent the sample names. `bam` is the file name of CLIP-seq bam. `condition` will be coerced to factor, whose first level will be treated as IP, and the second level as input.}

\item{signal.type}{character, indicate the type of feature signal wanted, support `TPM` for Transcripts Per Kilobase Million, `FPKM` for Fragments Per Kilobase Million (for paired-end reads) and Reads Per Kilobase Million (for single-end reads), and `raw.count` for raw.count read counts}

\item{FUN.aggregate}{function used for aggreating signals within `conditon`, default to mean.}

\item{output.prefix}{character, prefix of output filename(s), .feature will be added, used only if `output.feature.anno` or `output.feature.signal` is TRUE}

\item{cores}{integer, number of available workers, sent to `nthreads` of `featureCounts`}

\item{verbose}{logical, whether (`TRUE`) to echo progess}

\item{FUN.contrast}{function used for contrast signals across `conditon`, default to substraction.}

\item{output.feature.signal}{logical, whether to output a table of feature signals}

\item{minMQS;minOverlap;isPairedEnd;...}{parameters for `featureCounts()`. `minMQS` is default to 10, and `minOverlap` is default to 12 (25\% of the typical read length of eCLIP-seq (~50bp)), and `isPairedEnd` is default to `TRUE`. Note: If you use Illumina HiSeq 2000, set `strandSpecific = 2`.}
}
\value{
a `surfData` object.
}
\description{
Quantify signals on event features from eCLIP-seq bam files.
}
\references{
\url{https://www.rna-seqblog.com/rpkm-fpkm-and-tpm-clearly-explained/}
}
\keyword{signal}
