% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{getIsoAS}
\alias{getIsoAS}
\title{Get AS events for one isoform (decision tree)}
\usage{
getIsoAS(li, gm)
}
\arguments{
\item{li}{boolean layout of i-th isoform.}

\item{gm}{gene model.}
}
\description{
Get AS events for one isoform (decision tree)
}
