% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R, R/prepare.R
\name{getMultiTxGene}
\alias{getMultiTxGene}
\title{Find genes that contain multiple transcripts}
\usage{
getMultiTxGene(anno, min = 2, max = Inf)

getMultiTxGene(anno, min = 2, max = Inf)
}
\arguments{
\item{anno}{data.frame or GRanges, annotation.}

\item{min}{integer, minimum number of tx"s.}

\item{max}{integer, maximum number of tx"s.}
}
\value{
gene_id's

a character vector of gene identifiers that possess multiple transcripts.
}
\description{
Find genes that contain multiple transcripts.
}
