% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{getURL}
\alias{getURL}
\title{Get URL's for UCSC Genome Browser}
\usage{
getURL(gr, hub = NULL)
}
\arguments{
\item{gr}{a \code{GRanges} object}
}
\value{
a vector of URL's
}
\description{
Get URL's for UCSC Genome Browser
}
