% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ground2}
\alias{ground2}
\title{Object for ground surfaces}
\format{A cimg object with 256x256 pixles.}
\usage{
ground2
}
\description{
This object is the same of ground1 but have being rotated by 5 dg and then displaced 5 pixels to the right and 10 to the up.
}
\keyword{datasets}

