% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ground3}
\alias{ground3}
\title{Object for ground surfaces}
\format{A cimg object with 256x256 pixles.}
\usage{
ground3
}
\description{
This object is the same of ground1 but have being displaced 5 pixels to the right and 10 to the up followed by a rotation of 5 dg.
}
\keyword{datasets}

