% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{heatmapAUC}
\alias{heatmapAUC}
\alias{heatmapAUC,SummarizedExperiment-method}
\alias{heatmapAUC,daseqResults-method}
\title{Heatmap of AUC score}
\usage{
heatmapAUC(object, ...)

\S4method{heatmapAUC}{SummarizedExperiment}(object, group = NULL)

\S4method{heatmapAUC}{daseqResults}(object, group = NULL)
}
\arguments{
\item{object}{a \code{surf} object output by \link{daseq}.}

\item{...}{various parameters.}

\item{group}{an optional \code{factor} vector, group of rows.}
}
\value{
a \code{ggplot} object.
}
\description{
This function produces the heatmap of AUC scores.

Row blocking is allowed by providing the \code{group} parameter.
}
