% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{import}
\alias{import}
\title{Import annotation files}
\usage{
import(...)
}
\arguments{
\item{...}{whatever input to \link{rtracklayer::import}.}
}
\value{
a \code{GRanges} object.
}
\description{
This is a wrapper of \link{rtracklayer::import} with specialty in genome annotation.
Specifically, it checks whether "gene" or "transcript" exist in the annotation and fill them in if possible.
}
