% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faseq.R
\name{inferredFeature}
\alias{inferredFeature}
\title{Extract SURF-inferred location features}
\usage{
inferredFeature(object)
}
\arguments{
\item{object}{a \code{surf} object from \link{faseq} or \link{faseqInfer}.}
}
\value{
a \code{GRanges} object of all SURF-inferred location features.
}
\description{
Extract SURF-inferred location features
}
