% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associate.R
\name{inferTarget}
\alias{inferTarget}
\title{Extract genomic ranges inferred by SURF}
\usage{
inferTarget(sres, fdr.cutoff = 0.05, signal.cutoff = 20)
}
\arguments{
\item{sres}{SURF data and results}

\item{fdr.cutoff}{significance cutoff for the adjusted p-values of DrSeq.}

\item{signal.cutoff}{threshold cut-off for the eCLIP signals, default to 20.
Set this to 0 if dont wnat to filter those location with low eCLIP signals of the RBP.}
}
\value{
a GRanges object, all genomic ranges where the RBP is likely to interact and regulate some ATR events.
}
\description{
Extract genomic ranges inferred by SURF. a targeted feature is defined by (1) event (2) group (3) inferred (4) signal
}
