% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp.R
\name{interp}
\alias{interp}
\title{Fill the gaps using nearest neighbors}
\usage{
interp(im, cubic = FALSE, extrapolate = FALSE, npass = 3)
}
\arguments{
\item{im}{an cimg object with NAs}

\item{cubic}{if TRUE, use cubic interpolation. If FALSE, use linear (default FALSE)}

\item{extrapolate}{allow extrapolation (to values outside the image)}

\item{npass}{number of passes of the algorithm (usefull to fill larger holes)}
}
\value{
a \code{\link[imager]{cimg}} object.
}
\description{
This function fill the NA values of a cimg object by averaging neighbors
}
