% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artificial-data.R
\docType{data}
\name{LFS79.0910}
\alias{LFS79.0910}
\title{Artificial 1979 Sep-Oct LFS dataset}
\format{
A data frame with three columns: \code{y1} and \code{y2} describing
individual job status ( \code{E} = "Employed"; \code{U}; "Unemployed"; \code{N} = "Not in the labor force") in each month,
and a vector of selection probailities \code{probs}.
}
\usage{
data("LFS79.0910")
}
\description{
Artificial dataset based on Stasny (1987) LFS counts for the 1979 September-October flows.
}
\keyword{real}
