% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{list_rbind}
\alias{list_rbind}
\title{Bind list by row into data.frame}
\usage{
list_rbind(x, save.names = F)
}
\arguments{
\item{x}{list, all elements are vectors of the same length or array of the same column size}

\item{save.names}{logical or character, if not FALSE, save list names into an new column}
}
\description{
Bind by row a \code{list} of \code{data.frame} with the same \code{ncol} into \code{data.frame}. 
In particular, if the input is a list of vector-like objects (e.g. numeric, atomic, double, etc), each unit of list will be coerced into row vector.
In addition, the function allows to save list names if needed.
}
