% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cols.R
\name{make_cols}
\alias{make_cols}
\title{Add new columns to a data frame}
\usage{
make_cols(x, new.cols)
}
\arguments{
\item{x}{Data frame}

\item{new.cols}{Matrix or data frame of values of new columns.}
}
\description{
Useful for preparing data for facet plots, especially when combining many data sets or adding many columns
}
\examples{
x<-data.frame(rnorm(10))
names(x)<-'est'
new.cols<-cbind('Landings','Prediction')
colnames(new.cols)<-list('source','response')
make_cols(x,new.cols)
}
