% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_num_pretty.R
\name{make_num_pretty}
\alias{make_num_pretty}
\title{Make number pretty}
\usage{
make_num_pretty(x, r = 3)
}
\arguments{
\item{x}{A number}

\item{r}{Round, number of decimal places to use. Default is 3.}
}
\description{
Make number pretty
}
\examples{
make_num_pretty(12356789)
make_num_pretty(12356789,5)
make_num_pretty(0.12356789)
make_num_pretty(0.12356789,5)
make_num_pretty(8.1e-09)
make_num_pretty(8.1e-09,10)
}
