% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{mean.imlist}
\alias{mean.imlist}
\title{Custom mean for \code{imlist} objects}
\usage{
\method{mean}{imlist}(x, ...)
}
\arguments{
\item{x}{\code{imlist} object}

\item{...}{Not used on the current implementation.}
}
\value{
mean surface of all surfaces inside \code{imlist} object
}
\description{
Custom mean for \code{imlist} objects
}

