% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parseEvent}
\alias{parseEvent}
\title{Parse AS/ATS/APA events from genome annotation}
\usage{
parseEvent(
  annotation.file,
  annotation.format = "gtf",
  cores = max(1, detectCores() - 2),
  save.rds = F,
  rds.prefix = "",
  min.event.length = 6,
  depth.exon = 100,
  depth.intron = 300,
  remove.duplicate = T,
  verbose = T
)
}
\arguments{
\item{annotation.file}{diretory to the genome annotation.}

\item{annotation.format}{format of annotation file, default is GTF.}

\item{cores}{integer, number of workers available.}

\item{save.rds}{logical, whether (`TRUE`) to save intermediate result to .rds file. This is useful when running SURF in batch or large scale essay.}

\item{rds.prefix}{output prefix, default is empty and will output hidden files.}

\item{min.event.length}{numeric (positive), minimum length of a valid event.}

\item{depth.exon}{extended depth into exon, default 50 nt.}

\item{depth.intron}{extended depth into intron, default 300 nt.}

\item{remove.duplicate}{logical, whether (`TRUE`) to remove identical event duplicates (and keep one).}

\item{verbose}{logical, whether (`TRUE`) to print out progress.}
}
\description{
This function parses AS/ATS/APA events from genome annotation.
The task could be computationally demanding for unfiltered (raw) genome annotation.
}
