% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drseq.R
\name{plotDispFunc}
\alias{plotDispFunc}
\title{Plot Dispersion Functions}
\usage{
plotDispFunc(
  drr,
  x.limits = c(0.1, 10000),
  colrs = setNames(surf.colors, surf.events)
)
}
\arguments{
\item{drr}{a `drseqResults`` object.}
}
\value{
a `ggplot` object.
}
\description{
Plot Dispersion Functions
}
\details{
By default, ATR type are colored with the `Paired` palette (see \url{http://colorbrewer2.org}).
}
