% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired-data.R
\docType{data}
\name{pnadc1}
\alias{pnadc1}
\alias{pnadc2}
\alias{real}
\title{Real Datasets}
\format{
Two objects of class \code{data.frame}: \code{pnadc1, pnadc2}
}
\usage{
data("real")
}
\description{
North region subset of the Brazilian Continous Household Survey,
containing data from paired observations in the 1st and 2nd quarters of 2012.
}
\seealso{
\code{\link{df2}} for artificial data from a pseudo-population.
}
\keyword{real}
