% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{^.imlist}
\alias{^.imlist}
\title{Custom power of two \code{imlist} objects}
\usage{
\method{^}{imlist}(surf1, surf2)
}
\arguments{
\item{surf1}{\code{imlist} object}

\item{surf2}{\code{imlist} object}
}
\description{
Custom power of two \code{imlist} objects
}

