% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read.zip}
\alias{read.zip}
\title{Reads all surface '.txt' files inside a ".zip" file}
\usage{
read.zip(file, file.n = NULL, cutoff = c(0.001, 0.999), na.rm = TRUE)
}
\arguments{
\item{file}{A string withe the name of the '.txt' file to be loaded.}

\item{file.n}{the number(s) of file(s) to be read, default is NULL which read all files}

\item{cutoff}{cutoff}

\item{na.rm}{interpolation algorithm}
}
\value{
a \code{\link[imager]{imlist}} object.
}
\description{
This function reads a standard '.txt' file and converts it into a list of
\code{cimg} objects
}
\examples{
file <- system.file("extdata", "ground.zip", package = "surf")
surf <- read.zip(file)
par(mfrow = c(1,2))
plot(surf, layout = "row")
}
