% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-methods.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{coef}
\alias{vcov}
\alias{SE}
\alias{cv}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{stats}{\code{\link[stats]{coef}}, \code{\link[stats]{vcov}}}

  \item{survey}{\code{\link[survey]{cv}}, \code{\link[survey]{SE}}}
}}

