% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{rm.outliers}
\alias{rm.outliers}
\title{bla bla}
\usage{
rm.outliers(x, lower = 0.01, upper = 0.99)
}
\arguments{
\item{x}{array for input data}

\item{lower}{numeric value for lower quantile bound}

\item{upper}{numeric value for upper quantile bound}
}
\description{
bla bla
}
