% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired-data.R
\docType{data}
\name{s.qtr}
\alias{s.qtr}
\title{Brazilian Continous Household Survey Paired Data: 2st Quarter}
\format{An object of class \code{data.frame}.}
\usage{
s.qtr
}
\description{
Subset of the Brazilian Continous Household Survey for the Amazonas state,
containing data from paired observations in the 2nd quarter.
}
\seealso{
\code{\link{f.qtr}}
}
\keyword{datasets}
