% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_wsp.R
\name{save_wsp}
\alias{save_wsp}
\title{Save workspace with script name and today's date}
\usage{
save_wsp(path)
}
\arguments{
\item{path}{Path to save the workspace}
}
\description{
Must execute within script
}
