% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daseq.R
\name{scatterAUC}
\alias{scatterAUC}
\title{Scatter plot for DASeq results}
\usage{
scatterAUC(
  dar,
  alpha = 0.01,
  delta = 0.06,
  min.size = 50,
  remove.size = 10,
  force = 5,
  group = NULL,
  color = NULL
)
}
\arguments{
\item{dar}{a `daseqResults` object.}

\item{alpha}{numeric, FDR cut-off, default is 0.05.}

\item{delta}{numeric, lower cut-off for abs. diff. of AUC's between two conditions (default to 0.05).}

\item{min.size}{numeric, minimum size of the target size to color}

\item{remove.size}{numeric, minimum size of the target size to even display.}

\item{force}{Force of repulsion between overlapping text labels. Defaults
to 1.}
}
\value{
a `ggplot` object.
}
\description{
Scatter plot for DASeq results
}
