% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associate.R
\name{screenSurfEvent}
\alias{screenSurfEvent}
\title{Screen SURF training samples}
\usage{
screenSurfEvent(
  event,
  drseq.fdr = 0.05,
  read.length = 100,
  min.adjMean = 0.05,
  filter.overlap.event = T,
  verbose = F
)
}
\arguments{
\item{event}{a `drseqResults` object output by by `drseq()` or a `surfData` object.}

\item{drseq.fdr}{FDR (BH procedure) adjusted p-value cut-off.}

\item{read.length}{numeric, RNA-seq read length. Default is 100 bp (e.g., Illumina TruSeq). This is used to adjust event base count, which is then used to select the representative events if replicated.}

\item{min.adjMean}{numeric, adjusted event base mean threshold.}

\item{filter.overlap.event}{logical, whether (default to `TRUE`) to select one representitive event from overlapping ones and remove the others.}

\item{verbose}{logical, whether (default to `FALSE`) to print out basic summary statistics.}
}
\value{
a `surfData` object, with added one columns to `event`,
\item{adjMean}{adjusted base mean of the event from RNA-seq data.}
\item{group}{group labels of ATR events, `inc` for increased REU upon RBP knock-down, and `dec` for decreased, and `ctrl` for no-changed.}
\item{included}{logical, indicating whether the event is included into SURF ananlysis module 2.}
}
\description{
Extract response and covariates for the analysis module 2 of SURF.
}
