% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{sd.imlist}
\alias{sd.imlist}
\title{Custom sd for \code{imlist} objects}
\usage{
sd.imlist(surf)
}
\arguments{
\item{surf}{\code{imlist} object}
}
\value{
standard deviation surface of all surfaces inside \code{imlist} object
}
\description{
Custom sd for \code{imlist} objects
}

