% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfydesign.R
\name{sfydesign}
\alias{sfydesign}
\title{Flow estimation using complex surveys}
\usage{
sfydesign(
  ids,
  probs = NULL,
  strata = NULL,
  fpc = NULL,
  data.list = NULL,
  nest = FALSE,
  check.strata = !nest,
  weights = NULL
)
}
\arguments{
\item{ids}{Formula or data frame specifying cluster ids from largest level to smallest level, ~0 or ~1 is a formula for no clusters.}

\item{probs}{Formula or data frame specifying cluster sampling probabilities}

\item{strata}{Formula or vector specifying strata, use NULL for no strata}

\item{fpc}{Finite population correction: see Details in \link[survey]{svydesign}.}

\item{data.list}{list of data frames to look up variables in the formula arguments.}

\item{nest}{If TRUE, relabel cluster ids to enforce nesting within strata}

\item{check.strata}{If TRUE, check that clusters are nested in strata}

\item{weights}{Formula or vector specifying sampling weights as an alternative to prob. Notice: longitudinal weights.}
}
\value{
Object of class \code{surflow.design}, which are \link[survey]{svydesign} with
a \code{data.list} attribute containing the data for each survey round.
}
\description{
Specify a complex survey design for flow estimation.
}
\details{
The arguments of this function are those of \code{\link[survey]{svydesign}},
except for the \code{data.list}, which is a list of \code{data.frames}.

The first \code{data.frame} in \code{data.list} must contain the survey design information columns; i.e., clusters, strata, sampling probabilities, etc.
}
\examples{
# load data
data( "artificial" )

# create surflow design object
flowdes <-
sfydesign( ids = ~ 1 ,
               probs = ~ prob ,
               data.list = list( dfa0 , dfa1 ) ,
               nest = TRUE )

# describe object
summary( flowdes )

}
\references{
Gutierrez, A., Trujillo, L. \& Silva, N. (2014). The estimation of gross flows in complex surveys with random nonresponse,
Survey Methodology 40(2), pp. 285-321.

Lumley, Thomas S. (2010). Complex Surveys: A Guide to Analysis Using R. Wiley Publishing.
}
\seealso{
\code{\link[survey]{svydesign}}
}
\author{
Guilherme Jacob
}
\keyword{survey}
