% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_ggplot_legend.R
\name{shift_ggplot_legend}
\alias{shift_ggplot_legend}
\title{Shift the ggplot legend into an empty facet space}
\source{
https://stackoverflow.com/questions/54438495/shift-legend-into-empty-facets-of-a-faceted-plot-in-ggplot2
}
\usage{
shift_ggplot_legend(facet_plot)
}
\arguments{
\item{facet_plot}{A faceted ggplot}
}
\description{
Shift the ggplot legend into an empty facet space
}
\author{
Z.Lin, https://stackoverflow.com/users/8449629/z-lin
}
