% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset-method.R
\name{subset.survflow.design}
\alias{subset.survflow.design}
\title{Subset of survflow}
\usage{
\method{subset}{survflow.design}(x, subset, which.round = 0, ...)
}
\arguments{
\item{x}{a survflow design object}

\item{subset}{An expression specifying the subpopulation}

\item{which.round}{a vector of integers indicating which round to apply subsetting condition. Defaults to \code{which.round = 0}, and the expression is applied to the first round.}
}
\description{
Restrict a survey flow design to a subpopulation, keeping the original survey flow design information about number of clusters, strata.
}
\examples{
# load data
data( "initial" )
data( "final" )

# create survflow design object
flowdes <-
  svyflowdesign( ids = ~ upa ,
                 strata = ~ estrato ,
                 probs = ~ longprob ,
                 data.list = list( f.qtr , s.qtr ) ,
                 nest = TRUE )

# subset
flowdes <- subset( flowdes , v2007 == 1 )

# means
svymean( ~factor( vd4002 ) , flowdes , na.rm = TRUE ) # drops observations with missing in any rounds

}
