% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_by_group.R
\name{summarise_by_group}
\alias{summarise_by_group}
\title{Summarise by group}
\usage{
summarise_by_group(df, grp.cols, summarise.cols, fun)
}
\arguments{
\item{df}{A data frame}

\item{grp.cols}{Index the columns to group by. Can index by column name or column number.}

\item{summarise.cols}{Index the columns that should be summarised. Can index by column name or column number.}

\item{fun}{The function for summarising (e.g., mean, max etc.)}
}
\value{
A data frame
}
\description{
Summarise data frame variables according to group
}
\examples{
summarise_by_group(mtcars, "cyl", c(1, 3:ncol(mtcars)), max)
summarise_by_group(mtcars, 2, c(1, 3:ncol(mtcars)), max)
summarise_by_group(mtcars, c(2, 10), c(1, 3, 5:9), max)
summarise_by_group(mtcars, "cyl", c("mpg", "disp"), max)
summarise_by_group(mtcars, "cyl", "mpg", max)
summarise_by_group(mtcars, c("cyl",'gear'), c("mpg", "disp"), max)
summarise_by_group(mtcars, "cyl", c(1,3), max)
}
