% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super_res.R
\name{super_res}
\alias{super_res}
\title{High resolution image formation based on low resolution frames}
\usage{
super_res(LRf, uscale = 2, alpha = 1, ref_frame = NULL, step = NULL)
}
\arguments{
\item{LRf}{\code{imlist} object containing the Low Resolution frames}

\item{uscale}{numerical indicating the upscale factor}

\item{alpha}{numeric indicating the alpha parameter}

\item{ref_frame}{reference frame index.}

\item{step}{numeric indicating the step size in the gradient descent procedure}
}
\description{
bla bla
}
\examples{
\dontrun{
HR_set1 <- super_res(ground[(1:40)*2], uscale = 4, alpha = 5)
HR_set2 <- super_res(ground[(1:40)*2 - 1], uscale = 4, alpha = 5)
r <- register(HR_set1, HR_set2, method = 'taylor')
rHR_set1 <- transform(HR_set1, r[1], r[2], r[3])
mean(abs(HR_set1 - HR_set2))
mean(abs(rHR_set1 - HR_set2))
mean(sapply(lapply(ground[1:79] - ground[2:80], abs), mean))
par(mfrow = c(1,3))
plot(ground[[20]], axes = FALSE, interpolate = FALSE)
plot(rHR_set1, axes = FALSE, interpolate = FALSE)
plot(abs(rHR_set1 - HR_set2), axes = FALSE, interpolate = FALSE)
}
}
