% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{class}
\name{surf-class}
\alias{surf-class}
\alias{genePartsList,surf-method}
\alias{drseqResults,surf-method}
\alias{faseqResults,surf-method}
\alias{daseqResults,surf-method}
\title{SURF class}
\usage{
\S4method{genePartsList}{surf}(object)

\S4method{drseqResults}{surf}(object)

\S4method{faseqResults}{surf}(object)

\S4method{daseqResults}{surf}(object)
}
\arguments{
\item{object}{a \code{surf} object output by \link{daseq}.}
}
\description{
The \code{surf} class is an all-in-one analytic object used in SURF framework.
A \code{surf} object contains results of DrSeq (analysis module 1), FASeq (analysis module 2), and DASeq (discovery module 1).
In addition, it also contain gene (isoform) parts list, parsing from genome annotation.
The gene parts list is essential for ATR event construction.
If the any analysis module or discovery module is performed, the sample metadata will also be recorded.
Each of the components mentioned above can be accessed through a function listed below.
}
\references{
Chen, F., & Keles, S. (2020). SURF: integrative analysis of a compendium of RNA-seq and CLIP-seq datasets highlights complex governing of alternative transcriptional regulation by RNA-binding proteins. \emph{Genome Biology}, 21(1), 1-23.
}
\seealso{
\link{genePartsList}, \link{drseqResults}, \link{faseqResults}, \link{daseqResults}, \link{sampleData}.
}
