% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surf-package.R
\docType{package}
\name{surf-package}
\alias{surf}
\alias{surf-package}
\title{surf: surf: Several useful R functions}
\description{
An assortment of functions for different projects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JoleneSutton/surf}
  \item Report bugs at \url{https://github.com/JoleneSutton/surf/issues}
}

}
\author{
\strong{Maintainer}: Jolene Sutton \email{jolene.sutton@dfo-mpo.gc.ca}

}
\keyword{internal}
