% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associate.R
\name{surfData}
\alias{surfData}
\title{SURF dataset}
\usage{
surfData(
  event,
  drseq.fdr = 0.05,
  read.length = 100,
  min.adjMean = 0.05,
  filter.overlap.event = T,
  eCLIPsample,
  verbose = F,
  ...
)
}
\arguments{
\item{event}{a `drseqResults` object output by by `drseq()` or a `surfData` object.}

\item{drseq.fdr}{FDR (BH procedure) adjusted p-value cut-off.}

\item{read.length}{numeric, RNA-seq read length. Default is 100 bp (e.g., Illumina TruSeq). This is used to adjust event base count, which is then used to select the representative events if replicated.}

\item{min.adjMean}{numeric, adjusted event base mean threshold.}

\item{filter.overlap.event}{logical, whether (default to `TRUE`) to select one representitive event from overlapping ones and remove the others.}

\item{verbose}{logical, whether (default to `FALSE`) to print out basic summary statistics.}

\item{...}{additional parameters for `getFeatureSignal`.}
}
\value{
a `surfData` object.
}
\description{
Prepare SURF dataset.
Specifically, (i) select SURF event based on read coverage,
(ii) quantify feature signals using eCLIP-seq data.
}
