% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_summary.R
\name{surveysummary}
\alias{surveysummary}
\alias{svytotal.survflow.design}
\alias{svytotal}
\alias{svymean}
\alias{svymean.survflow.design}
\title{Summary statistics for repeated sample surveys}
\usage{
\method{svytotal}{survflow.design}(x, design, which.round = NULL, ...)

\method{svymean}{survflow.design}(x, design, which.round = NULL, ...)
}
\arguments{
\item{x}{A formula}

\item{design}{survflow.design object}

\item{which.round}{a vector of integers indicating which dataset to be updated. Defaults to \code{which.round = NULL}, and the expression is applied to all datasets.}

\item{...}{arguments to be passed to \code{\link[survey]{surveysummary}} function.}
}
\value{
Objects of class "svystat" or "svrepstat", which are vectors with a "var" attribute giving the variance and a "statistic" attribute giving the name of the statistic.

These objects have methods for vcov, SE, coef, confint, svycontrast.
}
\description{
Compute means and totals for data from complex surveys with repeated samples.
}
\details{
The arguments of this function are those of \code{\link[survey]{surveysummary}}.
}
\examples{
# load data
data( "initial" )
data( "final" )

# create survflow design object
flowdes <-
svyflowdesign( ids = ~ upa ,
               strata = ~ estrato ,
               probs = ~ longprob ,
               data.list = list( f.qtr , s.qtr ) ,
               nest = TRUE )

# describe object
summary( flowdes )

# load data
data( "initial" )
data( "final" )

# create survflow design object
flowdes <-
  svyflowdesign( ids = ~ upa ,
                 strata = ~ estrato ,
                 probs = ~ longprob ,
                 data.list = list( f.qtr , s.qtr ) ,
                 nest = TRUE )

# totals
svytotal( ~factor( vd4002 ) , flowdes , na.rm = TRUE ) # drops observations with missing in any rounds
svytotal( ~factor( vd4002 ) , flowdes , na.rm = TRUE , which.dataset = 0 ) # drops observations with missing in the first round
svytotal( ~factor( vd4002 ) , flowdes , na.rm = TRUE , which.dataset = 1 ) # drops observations with missing in the second round

# means
svymean( ~factor( vd4002 ) , flowdes , na.rm = TRUE ) # drops observations with missing in any rounds
svymean( ~factor( vd4002 ) , flowdes , na.rm = TRUE , which.dataset = 0 ) # drops observations with missing in the first round
svymean( ~factor( vd4002 ) , flowdes , na.rm = TRUE , which.dataset = 1 ) # drops observations with missing in the second round

}
\references{
Lumley, Thomas S. (2010). "Complex Surveys: A Guide to Analysis Using R". Wiley Publishing.
}
\seealso{
\code{\link[survey]{surveysummary}}

\code{\link[survey]{surveysummary}}
}
\author{
Guilherme Jacob
}
