% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survflowdesign.R
\name{svyflowdesign}
\alias{svyflowdesign}
\title{Flow estimation using complex surveys}
\usage{
svyflowdesign(
  ids,
  probs = NULL,
  strata = NULL,
  fpc = NULL,
  data.list = NULL,
  nest = FALSE,
  check.strata = !nest,
  weights = NULL
)
}
\arguments{
\item{ids}{Formula or data frame specifying cluster ids from largest level to smallest level, ~0 or ~1 is a formula for no clusters.}

\item{probs}{Formula or data frame specifying cluster sampling probabilities}

\item{strata}{Formula or vector specifying strata, use NULL for no strata}

\item{fpc}{Finite population correction: see Details below}

\item{data.list}{list of data frames to look up variables in the formula arguments.}

\item{nest}{If TRUE, relabel cluster ids to enforce nesting within strata}

\item{check.strata}{If TRUE, check that clusters are nested in strata}

\item{weights}{Formula or vector specifying sampling weights as an alternative to prob. Notice: longitudinal weights.}

\item{variables}{Formula or data frame specifying the variables measured in the survey. If NULL, the data argument is used.}
}
\value{
Object of class "\code{survflow.design}", which are \link[survey]{svydesign} based on the initial survey
with a \code{data.pairs} attribute containing the data for subsequent surveys.
}
\description{
Specify a complex survey design for flow estimation.
}
\details{
The arguments of this function are those of \code{\link[survey]{svydesign}},
except for the \code{data.list}, which is a list of \code{data.frames}.
}
\examples{
# load data
data( "initial" )
data( "final" )

# create survflow design object
flowdes <-
svyflowdesign( ids = ~ upa ,
               strata = ~ estrato ,
               probs = ~ longprob ,
               data.list = list( f.qtr , s.qtr ) ,
               nest = TRUE )

# describe object
summary( flowdes )

}
\references{
Gutierrez, A., Trujillo, L. \& Silva, N. (2014). The estimation of gross flows in complex surveys with random nonresponse,
Survey Methodology 40(2), pp. 285-321.

Lumley, Thomas S. (2010). Complex Surveys: A Guide to Analysis Using R. Wiley Publishing.
}
\seealso{
\code{\link[survey]{svydesign}}
}
\author{
Guilherme Jacob
}
\keyword{survey}
