% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associate.R
\name{trainSurf}
\alias{trainSurf}
\title{Learning one unit of SURF}
\usage{
trainSurf(data, min.size = 60, trim = 0.025)
}
\arguments{
\item{data}{data.frame, contains training data for one RBP and one event type}

\item{min.size}{integer, the minimum size of "reliable" training set, default to `60`.}
}
\description{
Train a GLM model for one RBPxAS combination. This is a helper function of testPF.
}
