% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform}
\alias{transform}
\title{Rigid body transformation of a surface}
\usage{
transform(surf, a, u, v, order = "RT")
}
\arguments{
\item{surf}{\code{cimg} object}

\item{a}{numerical indicating the angle of rotation (in degrees)}

\item{u, v}{numerical indicating the displacement in the x and y direction (in
pixels)}

\item{order}{character indicating the order of the transformation: "RT"
indicates rotate then translate and "TR" indicate translate than rotate.
Logical 1 or 0 are also accepted for optimization porpuses.}
}
\value{
a transformed \code{cimg} object
}
\description{
This function performs a rigid body transformation of a surface file
}
\examples{
par(mfrow = c(2,2), mar = c(2,2,1,1))
plot(ground1, asp = 1)
plot(transform(ground1, 5, 10, -10), asp = 1)
plot(transform(transform(ground1, 5, 10, -10), -5, -10, 10, "TR"))
plot(error(transform(transform(ground1, 5, 10, -10), -5, -10, 10, "TR"), ground1))
mean(error(transform(transform(ground1, 5, 10, -10), -5, -10, 10, "TR"), ground1))
}

