% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{tximport2}
\alias{tximport2}
\title{Customized tximport from `DESeq2`}
\usage{
tximport2(
  files,
  type = c("none", "kallisto", "salmon", "sailfish", "rsem"),
  txIn = TRUE,
  txOut = FALSE,
  countsFromAbundance = c("no", "scaledTPM", "lengthScaledTPM"),
  tx2gene = NULL,
  reader = read.delim,
  geneIdCol,
  txIdCol,
  abundanceCol,
  countsCol,
  lengthCol,
  importer,
  collatedFiles,
  ignoreTxVersion = FALSE,
  quiet = F
)
}
\arguments{
\item{files}{character, files to transcriptome quantification results}
}
\value{
a list of length four, named as `abundance`, `counts`, `length`, and `countsFromAbundance`
}
\description{
Customized tximport from `DESeq2`
}
