% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-method.R
\name{update.surflow.design}
\alias{update.surflow.design}
\title{Add variables to a surflow design}
\usage{
\method{update}{surflow.design}(object, ..., rounds = NULL)
}
\arguments{
\item{object}{a surflow design object}

\item{...}{Arguments tag=expr add a new variable tag computed by evaluating expr in the survey data}

\item{rounds}{a vector of integers indicating which dataset to be updated. Defaults to \code{rounds = NULL}, and the expression is applied to all datasets.}
}
\description{
Update the data variables in a surflow design, either with a formula for a new set of variables or with an expression for variables to be added.
}
\examples{
# load data
data( "artificial" )

# create surflow design object
flowdes <-
sfydesign( ids = ~ 1 ,
               probs = ~ prob ,
               data.list = list( dfa0 , dfa1 ) ,
               nest = TRUE )

# update
flowdes <-
     update( flowdes ,
             idstatus = factor( v0 ,
                                levels = 1:3 ,
                                labels = c( "A" , "B" , "C" ) ) )

# evaluate function
svyflow( ~idstatus , flowdes )

}
