% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-method.R
\name{update.survflow.design}
\alias{update.survflow.design}
\title{Add variables to a survflow design}
\usage{
\method{update}{survflow.design}(object, which.round = NULL, ...)
}
\arguments{
\item{object}{a survflow design object}

\item{which.round}{a vector of integers indicating which dataset to be updated. Defaults to \code{which.round = NULL}, and the expression is applied to all datasets.}

\item{...}{Arguments tag=expr add a new variable tag computed by evaluating expr in the survey data}
}
\description{
Update the data variables in a survflow design, either with a formula for a new set of variables or with an expression for variables to be added.
}
\examples{
# load data
data( "initial" )
data( "final" )

# create survflow design object
flowdes <-
  svyflowdesign( ids = ~ upa ,
                 strata = ~ estrato ,
                 probs = ~ longprob ,
                 data.list = list( f.qtr , s.qtr ) ,
                 nest = TRUE )

# update
flowdes <- update( flowdes , id_ocup = ifelse( vd4002 ==1 , "ocup" , "desocup" ) )

# means
svymean( ~factor( id_ocup ) , flowdes , na.rm = TRUE ) # drops observations with missing in any rounds

}
