% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write.imlist}
\alias{write.imlist}
\title{Write a list of standard surface '.txt' file}
\usage{
write.imlist(x, file = "data")
}
\arguments{
\item{x}{a \code{\link[imager]{cimg}} or a \code{\link[imager]{imlist}}
object.}

\item{file}{A string withe the name of the file to be writen}
}
\description{
This function reads a standard '.txt' file and converts it into a \code{cimg}
object
}
\examples{
file <- system.file("extdata", "ground.zip", package = "surf")
surf <- read.zip(file)
write.imlist(surf)
}
