% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{addGeneTx}
\alias{addGeneTx}
\title{Add identifiers of genes and transcripts
Fill up missing "gene" and "transcript" in genome annotation
If the genome annotation lacks "gene" and "transcript" entries, add to it.}
\usage{
addGeneTx(anno)
}
\arguments{
\item{anno}{a `GRanges` object of genome annotation, from `import`.}
}
\value{
a `GRanges` object with added `gene_id` and `transcript_id` columns.
}
\description{
Add identifiers of genes and transcripts
Fill up missing "gene" and "transcript" in genome annotation
If the genome annotation lacks "gene" and "transcript" entries, add to it.
}
