% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{aggregateAUCbyCondition}
\alias{aggregateAUCbyCondition}
\title{Aggregate AUC by sample condition}
\usage{
aggregateAUCbyCondition(object, sampleData, FUN.aggregate = "median")
}
\arguments{
\item{object}{a \code{SummarizedExperiment} output from \link{calculateAUC}}

\item{FUN.aggregate}{function, used for aggreating AUC within \code{conditon}, default to \code{mean}.}
}
\value{
a \code{data.frame} with 4 columns: \code{set}, \code{condition.1}, \code{condition.2}, and \code{diff}.
}
\description{
Aggregate AUC by sample condition
}
