% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate-design.R
\name{as.surfrdesign}
\alias{as.surfrdesign}
\title{Convert surflow design to use replicate weights}
\usage{
as.surfrdesign(design, ...)
}
\arguments{
\item{design}{Object of class surflow.design}

\item{...}{Arguments passed to \code{\link[survey]{as.svrepdesign}}}
}
\value{
Object of class \code{surflow.design}, \link[survey]{svrepdesign} with
a \code{data.list} attribute containing the data for each survey round.
}
\description{
Creates a replicate-weights survey flow design object from a traditional strata/cluster survey flow design object.
JK1 and JKn are jackknife methods, BRR is Balanced Repeated Replicates and Fay is Fay's modification of this,
bootstrap is Canty and Davison's bootstrap, subbootstrap is Rao and Wu's (n-1) bootstrap,
and mrbbootstrap is Preston's multistage rescaled bootstrap.
}
\details{
The arguments of this function are those of \code{\link[survey]{as.svrepdesign}},
}
\examples{
# load data
data( "artificial" )

# create surflow design object
flowdes <-
sfydesign( ids = ~ 1 ,
               probs = ~ prob ,
               data.list = list( dfa0 , dfa1 ) ,
               nest = TRUE )

# transform in replicate design
flowdes_rep <- as.surfrdesign( flowdes , type = "bootstrap" , replicates = 50 )

}
\references{
Gutierrez, A., Trujillo, L. \& Silva, N. (2014). The estimation of gross flows in complex surveys with random nonresponse,
Survey Methodology 40(2), pp. 285-321.

Lumley, Thomas S. (2010). Complex Surveys: A Guide to Analysis Using R. Wiley Publishing.
}
\seealso{
\code{\link[survey]{as.svrepdesign}}
}
\author{
Guilherme Jacob
}
\keyword{survey}
