% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{c_granges}
\alias{c_granges}
\title{Concatenate a list of GRanges}
\usage{
c_granges(grl, use.names = T, sep = ".", save.names = F)
}
\arguments{
\item{grl}{a \code{list} of \code{GRanges}.}

\item{use.names}{\code{logical}, whether (\code{TRUE}) to inherit list names.}

\item{sep}{\code{character}, separator between list names and GRanges names.}

\item{save.names}{\code{logical} or \code{character}, if not \code{FALSE}, save list names as an attribute (i.e., \code{mcols()}).}
}
\value{
a \code{GRanges} object
}
\description{
Concatenate a list of GRanges
}
