% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{calculateAUC}
\alias{calculateAUC}
\title{Calculate AUC}
\usage{
calculateAUC(set, rankings, cores = 1, verbose = F, ...)
}
\arguments{
\item{rankings}{'Rankings' created for this dataset with
\code{\link[AUCell]{AUCell_buildRankings}}.}

\item{cores}{integer, number of computing cores to use.}

\item{verbose}{Should the function show progress messages? (TRUE / FALSE)}
}
\value{
a \code{SummarizedExperiment} object
}
\description{
Calculate AUC
}
