% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{clusterByGroup}
\alias{clusterByGroup}
\title{Hierarchical clustering of rows by groups}
\usage{
clusterByGroup(x, group)
}
\arguments{
\item{x}{matrix}

\item{group}{group label of rows, disregard names, will be coerce to factor}
}
\value{
character vector of length \code{nrow(x)}, row names.
If \code{x} doesn't have row.names, this function names it sequentially (from 1).
}
\description{
Hierarchical clustering of rows by groups
}
