% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{daseq}
\alias{daseq}
\title{Differential activity analysis (DASeq)}
\usage{
daseq(
  event = NULL,
  rankings,
  sampleData,
  target.type = "transcript",
  targetSets = NULL,
  controlSets = NULL,
  n.sample = 1000,
  cores = max(1, detectCores() - 2),
  verbose = F,
  ...
)
}
\arguments{
\item{event}{a \code{surf} object from \link{faseqInfer} or \link{faseq}.}

\item{rankings}{a \code{SummarizedExperiment} object from \link{getRankings}.}

\item{sampleData}{\code{data.frame}, external samples, which contain \code{condition} column,
whose row names must match the column names of \code{rankings}.}

\item{target.type}{\code{character(1)}, either "transcript" or "gene".}

\item{targetSets}{a named \code{list} of \code{character}, the target IDs.}

\item{controlSets}{a named \code{list} of \code{character}, the control IDs.}

\item{n.sample}{\code{integer}, number of times of controlSet sampling, default to 1000.}

\item{cores}{\code{integer}, number of computing cores.}

\item{verbose}{\code{logical}, whether (\code{TRUE}) to echo progress.}

\item{...}{additional parameters for \link[AUCell:AUCell_calcAUC]{AUCell::AUCell_calcAUC}.}
}
\value{
a \code{daseqResults} object if \code{targetSets} was given,
or a \code{surf} object if \code{event} was give.
}
\description{
Detect differential activity using the AUC measure and RNA-seq quantification.
For more details about methodology, see Details.
This function can be used as part of SURF or as well a stand-along analysis.
For the former, input a \code{surf} object to \code{event}.
For the latter, input \code{targetSets} and optionally \code{controlSets}.
}
\references{
Chen, F., & Keles, S. (2020). SURF: integrative analysis of a compendium of RNA-seq and CLIP-seq datasets highlights complex governing of alternative transcriptional regulation by RNA-binding proteins. \emph{Genome Biology}, 21(1), 1-23.
}
