% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daseq.R
\name{daseq}
\alias{daseq}
\title{Differential activity (via AUC)}
\usage{
daseq(
  rankings,
  targetSets,
  controlSets = NULL,
  n.sample = 1000,
  cores = max(1, detectCores() - 2),
  verbose = F,
  ...
)
}
\value{
a daseqResults object, which contains a DataFrame and some testing configurations.
}
\description{
Detect differential activity using the AUC measure and RNA-seq quantification.
}
