% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\docType{class}
\name{daseqResults-class}
\alias{daseqResults-class}
\alias{sampleData,daseqResults-method}
\title{DASeq class}
\usage{
\S4method{sampleData}{daseqResults}(object)
}
\description{
\code{daseqResults} is a stand-alone object of DASeq (the discovery module of SURF) results.
}
