% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{drseq}
\alias{drseq}
\title{DrSeq}
\usage{
drseq(
  event,
  sampleData,
  design = ~sample + exon + condition:exon,
  remove.overlap.exon = F,
  anno.prefix = "drseq.annotation",
  anno.format = "gff2",
  minMQS = 10,
  isPairedEnd = T,
  ...,
  fullModel = design,
  reducedModel = ~sample + exon,
  fitExpToVar = "condition",
  drseq.fdr = 0.05,
  read.length = 100,
  min.adjMean = 0.05,
  filter.overlap.event = T,
  cores = max(1, detectCores() - 2),
  verbose = F
)
}
\arguments{
\item{event}{a \code{surf} object from \link{parseEvent}.}

\item{sampleData}{\code{data.frame}, describes the RNA-seq samples and
contains at least two columns -- \code{bam} and \code{condition},
whose \code{row.names} represent sample names.}

\item{design}{
    A formula which specifies the design of the experiment. It must
    specify an interaction term between a variable from the sampleData
    columns with the 'exon' variable. By default, the design 
    will be '~ sample + exon + condition:exon'. This formula indicates
    the contrast between 'condition' and exon', i.e. differences in exon 
    usage due to changes in the 'condition' variable.  See the vignette 
    for more examples of other designs.
  }

\item{remove.overlap.exon}{\code{logical}, remove overlapping exons across genes (default to \code{FALSE}).}

\item{anno.prefix}{\code{character}, file names for outputting annotation files.
If prefix is absent, hidden file will be output to the current working directory.}

\item{anno.format}{\code{character}, e.g. "gtf", as accepted by \link[rtracklayer:export]{rtracklayer::export}.}

\item{minMQS}{as defined in \link[Rsubread:featureCounts]{Rsubread::featureCounts}.
Note that the default is customized for SURF (see details for more information).}

\item{isPairedEnd}{as defined in \link[Rsubread:featureCounts]{Rsubread::featureCounts}.
Note that the default is customized for SURF (see details for more information).}

\item{...}{parameters for \link[Rsubread:featureCounts]{Rsubread::featureCounts}.}

\item{fullModel}{
     The full model formula
  }

\item{reducedModel}{
     Null model formula.
  }

\item{fitExpToVar}{
    A variable name contained in the sample data. The expression 
    values will be fitted to this variable using the the formula 
    " ~ sample + fitExpToVar * exon".
  }

\item{drseq.fdr}{\code{numeric}, FDR (BH procedure) adjusted p-value cut-off.}

\item{read.length}{\code{numeric}, RNA-seq read length. Default is 100 bp (e.g., Illumina TruSeq). This is used to adjust event base count, which is then used to select the representative events if replicated.}

\item{min.adjMean}{\code{numeric}, adjusted event base mean threshold.}

\item{filter.overlap.event}{\code{logical}, whether (default to \code{TRUE}) to select one representitive event from overlapping ones and remove the others.}

\item{cores}{\code{integer}, number of available workers, sent to \code{nthreads} of \code{featureCounts}.}

\item{verbose}{\code{logical}, whether (\code{TRUE}) to echo progress.}
}
\value{
a \code{surf} object containing the DrSeq result in the \code{drseqResults} slot.
}
\description{
Perform the differential REU (DrSeq) test in a single command.
This function is a wrapper that calls the necessary functions in order for DrSeq.
}
\references{
Chen, F., & Keles, S. (2020). SURF: integrative analysis of a compendium of RNA-seq and CLIP-seq datasets highlights complex governing of alternative transcriptional regulation by RNA-binding proteins. \emph{Genome Biology}, 21(1), 1-23.
}
