% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drseq.R
\name{drseqData}
\alias{drseqData}
\title{Create DrSeq dataset}
\usage{
drseqData(
  anno,
  anno_event,
  sampleData,
  design = ~sample + exon + condition:exon,
  remove.overlap.exon = F,
  create.annotation = T,
  anno.prefix = "drseq.annotation",
  anno.format = "gff2",
  export.count = F,
  count.prefix = "drseq.count",
  minMQS = 10,
  isPairedEnd = T,
  cores = max(1, detectCores() - 2),
  verbose = F,
  ...
)
}
\arguments{
\item{anno}{GRanges, genome annotation}

\item{anno_event}{GRanges, AS/ATS/APA event annotation}

\item{sampleData}{data.frame, describes the RNA-seq samples, contains at least two columns -- `bam` and `condition`, the row.names represent sample names}

\item{design}{data.frame, the design of RNA-seq experiment}

\item{remove.overlap.exon}{logical, remove overlaping exons across genes (default to FALSE)}

\item{create.annotation}{logical, whether (`TRUE`) to create a new annotation file for DrSeq. This option is useful when applying DrSeq to a large number of assays, where annotation can be shared globally once generated.}

\item{anno.prefix}{file names for outputting annotation files}

\item{export.count}{logical, whether the count matrix should be export to a file}

\item{cores}{integer, number of available workers, sent to `nthreads` of `featureCounts`}

\item{verbose}{logical, whether (`TRUE`) to echo progess}

\item{...}{additional parameters for `Rsubread::featureCounts`.}

\item{minMQS;isPairedEnd}{as defined in `featureCounts`, the default is customized for SURF.}
}
\value{
a `DEXSeqDataSet` object
}
\description{
Create a `DEXSeqDataSet` object for DrSeq, so as to use the `DEXSeq` as the engine of parameter estimation.
It is possible to output event read counts for later usage.
}
\details{
If you used Illumina HiSeq 2000, set `strandSpecific = 2` (reversed strand).
}
