% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drseq.R
\name{drseqFit}
\alias{drseqFit}
\title{DrSeq Fit}
\usage{
drseqFit(
  drd,
  fullModel = design(object),
  reducedModel = ~sample + exon,
  fitExpToVar = "condition",
  cores = max(1, detectCores() - 2),
  verbose = F
)
}
\arguments{
\item{drd}{a [surf] object from `drseqData`.}

\item{cores}{integer, number of computing workers.}

\item{verbose}{logical, whether (`TRUE`) to echo progess.}
}
\value{
a [surf] object with (1) `drseqResults` and `sampleData` slot updated and (2) three added columns:
\item{eventBaseMean}{base read coverage of the event from RNA-seq data.}
\item{padj}{adjusted p-value for differential REU.}
\item{logFoldChange}{estimated log2 fold change of REU: log2(condition of the 1st sample / another condition)}
}
\description{
Fit DrSeq models using the DEXSeq as the estimation engine.
}
\details{
the drseqResults slot contains extensive DrSeq results. To access the object, use [deseqResults()] function.
}
