% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{drseqResults}
\alias{drseqResults}
\title{DrSeq Results}
\usage{
drseqResults(object)
}
\arguments{
\item{object}{a \code{surf} object output by \link{drseq} or \link{drseqFit} or \link{drseqFilter}.}
}
\value{
a \code{drseqResults} object
}
\description{
DrSeq Results
}
