% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{error}
\alias{error}
\title{difference between two \code{cimg} objects}
\usage{
error(surf1, surf2)
}
\arguments{
\item{surf1}{\code{cimg} object}

\item{surf2}{\code{cimg} object}
}
\description{
difference between two \code{cimg} objects
}
\examples{
file <- system.file("extdata", "sur1.txt", package = "surf")
sur1 <- read.surf(file)
file <- system.file("extdata", "sur2.txt", package = "surf")
sur2 <- read.surf(file)
plot(error(sur1, sur2)) #melhorar exemplo
}

