% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{featureCounts}
\alias{featureCounts}
\title{Customized `Rsubread::featureCounts`}
\usage{
featureCounts(..., verbose = F)
}
\arguments{
\item{...}{See [Rsubread::featureCounts()].}
}
\value{
See [Rsubread::featureCounts()].
}
\description{
Wrapper of `Rsubread::featureCounts()` allowing verbose suppression
}
\details{
This redirects the output to `/dev/null`, so it assumes a UNIX-like
system.
}
