% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{featureCounts}
\alias{featureCounts}
\title{Customized \code{featureCounts}}
\usage{
featureCounts(..., verbose = F)
}
\arguments{
\item{...}{parameters for \code{Rsubread::featureCounts}.}
}
\value{
See \code{Rsubread::featureCounts} documentation.
}
\description{
This is a wrapper of \code{Rsubread::featureCounts} allowing verbose suppression.
This redirects the output to `/dev/null`, so it assumes a UNIX-like
system.
}
