% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form.R
\name{form}
\alias{form}
\alias{form.cimg}
\alias{form.imlist}
\title{Form removal}
\usage{
form(surf, form = "paraboloid")

\method{form}{cimg}(surf, form = "paraboloid")

\method{form}{imlist}(surf, form = "paraboloid")
}
\arguments{
\item{surf}{a \code{\link[imager]{cimg}} or a \code{\link[imager]{imlist}}
object.}

\item{form}{an optional parameter that overides the information given on
formula. Options are: "ellipsoid", "paraboloid". #enable partial match}
}
\value{
a corrected \code{\link[imager]{cimg}} or a
  \code{\link[imager]{imlist}} object.
}
\description{
Removes the underlying format of a surface.
}
\examples{
file <- system.file("extdata", "form.txt", package = "surf")
surf <- read.surf(file)
par(mfrow = c(1,3))
plot(surf, asp = 1)
plot(form(surf, form = 'paraboloid'), asp = 1)
plot(form(surf, form = 'ellipsoid'), asp = 1)

file <- system.file("extdata", "form.zip", package = "surf")
surf <- read.zip(file)
plot(form(surf, form = 'paraboloid'))
}

