% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form.R
\name{form}
\alias{form}
\alias{form.cimg}
\alias{form.imlist}
\title{Title}
\usage{
form(surf, form = "paraboloid", formula = NULL)

\method{form}{cimg}(surf, form = "paraboloid", formula = NULL)

\method{form}{imlist}(surf, form = "paraboloid", formula = NULL)
}
\arguments{
\item{surf}{a \code{\link[imager]{cimg}} or a \code{\link[imager]{imlist}} object.}

\item{form}{an optional parameter that overides the information given on formula. Options are: "eliptical", "paraboloid".}

\item{formula}{an optional object of class formula to pass to the lm function.}
}
\value{
a corrected \code{\link[imager]{cimg}} or a \code{\link[imager]{imlist}} object.
}
\description{
description
}
\examples{
file <- system.file("extdata", "form.txt", package = "surf")
surf <- read.surf(file)
par(mfrow = c(1,2))
plot(surf, asp = 1)
plot(form(surf), asp = 1)

}

