% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{getASName}
\alias{getASName}
\title{ATR event name}
\usage{
getASName(plas)
}
\arguments{
\item{plas}{a plas object containing \code{asName}}
}
\value{
a named (by event_id) vector of event type
}
\description{
ATR event name
}
