% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{getASName}
\alias{getASName}
\title{ATR event name}
\usage{
getASName(plas)

getASName(plas)
}
\arguments{
\item{plas}{an element of isoPLas object, containing `asName`}
}
\description{
get a named (by event_id) vector of event type for one gene

Retrieve a named (by event_id) vector of event type for one gene. This is a helper function to `getFeature()`.
}
