% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{getAUC}
\alias{getAUC}
\alias{getAUC,SummarizedExperiment-method}
\alias{getAUC,daseqResults-method}
\alias{getAUC,surf-method}
\title{Get AUC}
\usage{
getAUC(object)

\S4method{getAUC}{SummarizedExperiment}(object)

\S4method{getAUC}{daseqResults}(object)

\S4method{getAUC}{surf}(object)
}
\arguments{
\item{object}{a \code{surf} object output by \link{daseq}.}
}
\value{
a \code{matrix} of AUC, whose rows correspond to target sets and columns correspond to samples.
}
\description{
Get AUC measure for each target set (row) in every sample (column).
}
