% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{getEvent}
\alias{getEvent}
\title{Enumerate AS events
This will not merge SE events across consecutive exon
Find AS events from isoform parts list}
\usage{
getEvent(isoPL, cores = max(1, detectCores() - 2))
}
\arguments{
\item{isoPL}{isoform parts \code{list}}

\item{cores}{\code{integer}, number of computing cores to use.}
}
\value{
a \code{list} as input with added two components:
\item{asNum}{integer \code{matrix}, dim = dim of segment 0 means no event, and other integers indicate event number.}
\item{asName}{a \code{list} of named \code{character} vector, with each element corresponding to one isoform. name is event number, character is event name}
}
\description{
Enumerate AS events
This will not merge SE events across consecutive exon
Find AS events from isoform parts list
}
