% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{getIsoPartsList}
\alias{getIsoPartsList}
\title{Parse isoform parts list from annotation}
\usage{
getIsoPartsList(
  anno,
  gene_id = getMultiTxGene(anno),
  depth.intron = 300,
  cores = max(1, detectCores() - 2)
)
}
\arguments{
\item{anno}{a \code{GRanges} object of genome annotation, return by \link{import}.}

\item{gene_id}{\code{character}, gene_id"s to analyze, default to all multi-transcript genes.}

\item{depth.intron}{\code{integer}, extended depth into gene's flanks, default 300 nt.}

\item{cores}{\code{integer}, number of computing cores to use.}
}
\value{
a \code{list} named by gene_id, each of which contains:
\item{segment}{GRanges, genomic data of parts list.}
\item{label}{integer vector, gene model (0 for intron) and exon numbers (coded as integer 1,2,...).}
\item{layout}{lgCMatrix, transcript structure.}
}
\description{
Parse the isoform parts list from annotation.
}
