% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R, R/utility.R
\name{getMultiTxGene}
\alias{getMultiTxGene}
\title{Find multi-transciption gene}
\usage{
getMultiTxGene(anno, min = 2, max = Inf)

getMultiTxGene(anno, min = 2, max = Inf)
}
\arguments{
\item{anno}{data.frame or GRanges, annotation}

\item{min}{integer, minimum number of tx's}

\item{max}{integer, maximum number of tx's}
}
\value{
a character vector of gene identifiers that possess multiple transcripts.

gene_id's
}
\description{
Find genes that contain multiple transcripts.
}
