% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R, R/prepare.R
\name{getMultiTxGene}
\alias{getMultiTxGene}
\title{Find genes that contain multiple transcripts}
\usage{
getMultiTxGene(anno, min = 2, max = Inf)

getMultiTxGene(anno, min = 2, max = Inf)
}
\arguments{
\item{anno}{\code{data.frame} or \code{GRanges}, genome annotation.}

\item{min}{\code{integer}, minimum number of transcripts.}

\item{max}{\code{integer}, maximum number of transcripts..}
}
\value{
a \code{character} vector of gene_id's

a \code{character} vector of gene identifiers that possess multiple transcripts.
}
\description{
Find genes that contain multiple transcripts.
}
