% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daseq.R
\name{getTargetSet}
\alias{getTargetSet}
\title{Get target set based on \code{inferredFeature}}
\usage{
getTargetSet(event, id_column = "transcript_id", verbose = F)
}
\value{
list of character, sets of target identifiers.
}
\description{
Get target set based on \code{inferredFeature}
}
