% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daseq.R
\name{heatmapAUC}
\alias{heatmapAUC}
\title{Heatmap for DASeq results}
\usage{
heatmapAUC(dar, group = NULL)
}
\arguments{
\item{dar}{a `daseqResults` object.}

\item{group}{an optional `factor` vector, group of `targetSet`.}
}
\value{
a `ggplot` object.
}
\description{
Heatmap for DASeq results
}
