% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{import}
\alias{import}
\title{Perform `rtracklayer::import()`, followed by `addGeneTx()`}
\usage{
import(...)
}
\arguments{
\item{...}{whatever input to `rtracklayer::import`}
}
\value{
a GRanges object as a result of `addGeneTx(rtracklayer::import(...))`
}
\description{
Perform `rtracklayer::import()`, followed by `addGeneTx()`
}
