% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leveling.R
\name{level}
\alias{level}
\alias{level.cimg}
\alias{level.imlist}
\title{Level a surface}
\usage{
level(surf, method = "rot")

\method{level}{cimg}(surf, method = "rot")

\method{level}{imlist}(surf, method = "rot")
}
\arguments{
\item{surf}{a \code{\link[imager]{cimg}} or a \code{\link[imager]{imlist}}
object.}

\item{method}{one of two options: \code{'rot'} or \code{'sub'}.}
}
\value{
a corrected \code{\link[imager]{cimg}} or a
  \code{\link[imager]{imlist}} object.
}
\description{
Level the surface by assigning the horizontal reference based on a tilted
plane regression
}
\examples{
file <- system.file("extdata", "ground.txt", package = "surf")
surf <- read.surf(file)
par(mfrow = c(1,2))
plot(surf, asp = 1)
plot(level(surf), asp = 1)

file <- system.file("extdata", "ground.zip", package = "surf")
surf <- read.zip(file)
plot(surf, asp = 1)
plot(level(surf), asp = 1)
plot(imager::imlist(c(surf,level(surf))), asp = 1)
}

