% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{numberAS}
\alias{numberAS}
\title{Enumerate AS events}
\usage{
numberAS(as, gm)
}
\arguments{
\item{as}{a \code{factor} vector of AS events /isoform.}

\item{gm}{gene model.}
}
\value{
a \code{integer} vector of the same length as \code{as}, indexing the order of
ATR events.
}
\description{
Enumerate AS events
}
\details{
AFE, IAP, and consecutive SE events will be merged.
}
