% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{numberAS}
\alias{numberAS}
\title{Enumerate AS events}
\usage{
numberAS(as, gm)
}
\arguments{
\item{as}{factor of AS events /isoform}
}
\description{
Enumerate AS events
}
\details{
AFE, IAP, and consecutive SE events will be merged.
}
