% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{plotDispFunc}
\alias{plotDispFunc}
\alias{plotDispFunc,drseqResults-method}
\alias{plotDispFunc,surf-method}
\title{Plot Dispersion Functions}
\usage{
plotDispFunc(object, ...)

\S4method{plotDispFunc}{drseqResults}(object, x.limits = c(0.1, 10000))

\S4method{plotDispFunc}{surf}(object, ...)
}
\arguments{
\item{object}{a \code{drseqResults} object.}

\item{x.limits}{\code{numeric(2)}, limits for x-axis.}
}
\value{
a \code{ggplot} object.
}
\description{
Plot the fitted mean-dispersion functions of DrSeq models.
DrSeq fits up to eight mean-dispersion functions; each corresponds to one ATR event type.
This allows DrSeq to better account for the nuance in the over-dispersion presented by different ATR event types.
For more details, please refer to SURF paper.
}
\details{
By SURF default, ATR type are colored with the \code{Paired} palette (see \url{http://colorbrewer2.org}).
}
