% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drseq.R
\name{prepDrseqAnno}
\alias{prepDrseqAnno}
\title{Prepare DrSeq Annotation files}
\usage{
prepDrseqAnno(
  anno,
  anno_event,
  anno.prefix = "annotation.drseq",
  anno.format = "gff2",
  remove.overlap.exon = F,
  verbose = T
)
}
\arguments{
\item{anno}{GRanges, genome annotation}

\item{anno.prefix}{prefix of annotation file}

\item{anno.format}{character, e.g. "gtf", as defined by `rtracklayer::export()`}

\item{remove.overlap.exon}{- logical, remove overlaping exons across genes (default to FALSE)}

\item{verbose}{logical, whether (`TRUE`) to echo progess}

\item{event}{GRanges, AS/ATS/APA event}
}
\description{
To use the "useMetaFeatures" functionality of `Rsubread`, we need the GTF input. This function is a helper function that produce the GTF files needed.
}
