% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{prepDrseqAnno}
\alias{prepDrseqAnno}
\title{Prepare DrSeq annotation files}
\usage{
prepDrseqAnno(
  anno_event,
  anno.prefix = "annotation.drseq",
  anno.format = "gff2",
  remove.overlap.exon = F,
  cores = max(1, detectCores() - 2),
  verbose = T
)
}
\arguments{
\item{anno_event}{a \code{surf} object}

\item{anno.prefix}{\code{character}, prefix of exon/event annotation files for saving. These files are needed by \link[Rsubread:featureCounts]{Rsubread::featureCounts}.}

\item{anno.format}{\code{character}, e.g. "gtf", as accepted by \link[rtracklayer:export]{rtracklayer::export}.}

\item{remove.overlap.exon}{\code{logical}, remove overlapping exons across genes (default to \code{FALSE}).}

\item{cores}{\code{integer}, number of available workers, sent to \code{nthreads} of \link[Rsubread:featureCounts]{Rsubread::featureCounts}.}

\item{verbose}{\code{logical}, whether (\code{TRUE}) to echo progress.}
}
\value{
\code{NULL}, the function is a procedure and only output/export results to file system, except for messages and warnings.
}
\description{
To use the \code{useMetaFeatures} functionality of \link[Rsubread:featureCounts]{Rsubread::featureCounts}, we need the GTF input.
This function helps to produce the \emph{two} annotation files ("event" and "exon") needed by \link{drseqCount}.
}
