% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read.surf}
\alias{read.surf}
\title{Reads standard surface '.txt' file}
\usage{
read.surf(file, res = NULL)
}
\arguments{
\item{file}{A string withe the name of the '.txt' file to be loaded.}

\item{res}{Resolution for the input data, if \code{NULL} the algorithm will guess its value}
}
\value{
a \code{\link[imager]{cimg}} object.
}
\description{
This function reads a standard '.txt' file and converts it into a \code{cimg} object
}
\examples{
file <- system.file("extdata", "ground.txt", package = "surf")
surf <- read.surf(file)
plot(surf, asp = 1)
}

