% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.surf.R
\name{read.zip}
\alias{read.zip}
\title{Reads all surface '.txt' files inside a ".zip" file}
\usage{
read.zip(file, res = NULL)
}
\arguments{
\item{file}{A string withe the name of the '.txt' file to be loaded.}

\item{res}{Resolution for the input data, if \code{NULL} the algorithm will guess its value}
}
\value{
a \code{\link[imager]{imlist}} object.
}
\description{
This function reads a standard '.txt' file and converts it into a list of \code{cimg} objects
}
\examples{
file <- system.file("extdata", "ground.zip", package = "surf")
surf <- read.zip(file)
par(mfrow = c(1,2))
lapply(surf, plot, asp = 1)
}

