% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relocate.R
\name{relocate}
\alias{relocate}
\title{Relocation of a surface}
\usage{
relocate(surf, targ, control = NULL)
}
\arguments{
\item{surf}{\code{cimg} object}

\item{targ}{\code{cimg} object}

\item{control}{for the optimization algorithm}
}
\value{
a relocated \code{cimg} object
}
\description{
This function relocates a surface to match some reference sample.
}
\examples{
system.time(r_ground2 <- relocate(ground2, ground1,
  control = list(max.generations = 5, pop.size = 50, BFGS = FALSE)))
  # for better results, increas population size, max.generations, and enable BFGS
par(mfrow = c(2,2), mar = c(2,2,1,1))
plot(ground1, asp = 1)
plot(ground2, asp = 1)
plot(r_ground2, asp = 1)
plot(error(ground1, r_ground2), asp = 1)
mean(error(ground1, r_ground2))
}

