% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{sampleData}
\alias{sampleData}
\alias{sampleData,SummarizedExperiment-method}
\title{Sample Data}
\usage{
sampleData(object, ...)

\S4method{sampleData}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{any object that contains a \code{sampleData} slot.}

\item{...}{various parameters.}
}
\value{
\code{DataFrame} or \code{DataFrameList} (if \code{length(type)>1}).
}
\description{
Sample Data
}
