% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{sampleData}
\alias{sampleData}
\title{Sample Data}
\usage{
sampleData(object, ...)
}
\arguments{
\item{type}{\code{character}, any subset of "RNA-seq", "CLIP-seq", and "External".}
}
\value{
\code{DataFrame} or \code{DataFrameList} (if \code{length(type)>1}).
}
\description{
Sample Data
}
