% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfydesign.R
\name{sfydesign}
\alias{sfydesign}
\title{Flow estimation using complex surveys}
\usage{
sfydesign( ids, probs=NULL, strata=NULL, fpc=NULL, data=NULL, nest=FALSE,
check.strata = !nest, weights=NULL, pps=FALSE,...)
}
\arguments{
\item{ids}{Formula or data frame specifying cluster ids from largest level to smallest level, ~0 or ~1 is a formula for no clusters.}

\item{probs}{Formula or data frame specifying cluster sampling probabilities}

\item{strata}{Formula or vector specifying strata, use NULL for no strata}

\item{fpc}{Finite population correction: see Details in \link[survey]{svydesign}.}

\item{data}{list of data frames to look up variables in the formula arguments.}

\item{nest}{If TRUE, relabel cluster ids to enforce nesting within strata}

\item{check.strata}{If TRUE, check that clusters are nested in strata}

\item{weights}{Formula or vector specifying sampling weights as an alternative to prob. Notice: longitudinal weights.}

\item{pps}{\code{"brewer"} to use Brewer's approximation for PPS sampling without replacement.
\code{"overton"} to use Overton's approximation. An object of class \link[survey]{HR} to use the Hartley-Rao approximation.
An object of class \link[survey]{ppsmat} to use the Horvitz-Thompson estimator.}

\item{...}{for future expansion. See \link[survey]{svydesign} for more information.}
}
\value{
Object of class \code{surflow.design}, which are \link[survey]{svydesign} with
a \code{data} attribute containing the data for each survey round.
}
\description{
Specify a complex survey design for flow estimation.
}
\details{
The arguments of this function are those of \code{\link[survey]{svydesign}},
except for the \code{data}, which, in this case, is a list of \code{data.frames} with \emph{paired observations}; i.e., each
row must refer to the same individual across datasets.

The first \code{data.frame} in \code{data} must contain the survey design information columns; i.e., clusters, strata, sampling probabilities, etc.
}
\examples{
# load data
data( "artificial" )

# create surflow design object
flowdes <-
sfydesign( ids = ~ 1 ,
               probs = ~ prob ,
               data = list( df0 , df1 ) ,
               nest = TRUE )

# describe object
summary( flowdes )

}
\references{
ROJAS, H. A. G.; TRUJILLO, L.; SILVA, P. L. N. The estimation of gross flows in complex surveys with random nonresponse.
\emph{Survey Methodology}, v. 40, n. 2, p. 285–321, dec. 2014. URL \url{https://www150.statcan.gc.ca/n1/en/catalogue/12-001-X201400214113}.

LUMLEY, T. \emph{Complex Surveys:} A guide to analysis using R.
Hoboken: John Wiley & Sons, 2010. (Wiley Series in Survey Methodology). ISBN 978-0-470-28430-8.
}
\seealso{
\code{\link[survey]{svydesign}}
}
\author{
Guilherme Jacob
}
\keyword{survey}
