% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{stream}
\alias{stream}
\title{Reconstruct exon parts list to isoform parts list}
\usage{
stream(ex, gm)
}
\arguments{
\item{ex}{exon parts list.}

\item{gm}{gene model.}
}
\description{
This is a helper function of \link{getEvent}.
}
