% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{stream}
\alias{stream}
\title{Reconstruct exon parts list to isoform parts list}
\usage{
stream(ex, gm)
}
\arguments{
\item{ex}{exon parts list.}

\item{gm}{gene model.}
}
\value{
a \code{logical} vector of the same length of \code{gm}, indicating exonic
(\code{1}) or intronic (\code{0}) segments.
}
\description{
This is a helper function of \link{getEvent}.
}
