% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset-method.R
\name{subset.surflow.design}
\alias{subset.surflow.design}
\title{Subset of surflow}
\usage{
\method{subset}{surflow.design}(x, subset, rounds = NULL, ..., on.full.design = FALSE)
}
\arguments{
\item{x}{a surflow design object}

\item{subset}{An expression specifying the subpopulation}

\item{rounds}{a vector of integers indicating which round to apply subsetting condition.
Defaults to \code{rounds = NULL}, and the expression is applied to all rounds, keeping observation that were true in at least of the rounds.}

\item{...}{future expansion}

\item{on.full.design}{A logical, indicating wether the subset should be applied to full design weights. Only meaningful for endogenous categories.}
}
\description{
Restrict a survey flow design to a subpopulation, keeping the original survey flow design information about number of clusters, strata.
}
\examples{
# load data
data( "real" )

# create surflow design object
flowdes <-
  sfydesign( ids = ~ upa ,
             probs = ~ v1027 ,
             strata = ~ estrato ,
             data = list( pnadc1 , pnadc2 ) ,
             nest = TRUE )

# subset
flowdes <- subset( flowdes , v2009 >= 14 )

}
