% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svyflow.R
\name{svyflow}
\alias{svyflow}
\alias{svyflow.survey.design2}
\alias{svyflow.svyrep.design}
\alias{svyflow.surflow.design}
\title{Summary statistics for repeated sample surveys}
\usage{
\method{svyflow}{survey.design2}(x, design, flow.type, rounds, max.iter, na.rm, ...)

\method{svyflow}{svyrep.design}(x, design, flow.type, rounds, max.iter, na.rm, ...)

\method{svyflow}{surflow.design}(
  x,
  design,
  flow.type = NULL,
  rounds = c(0, 1),
  max.iter = 10,
  na.rm = FALSE,
  ...
)
}
\arguments{
\item{x}{A formula}

\item{design}{surflow.design object}

\item{flow.type}{type of flow to estimate: "gross" for counts and "net" for probabilities. Defaults to \code{flow.type = "gross"}.}

\item{rounds}{a vector of integers indicating which round to use. Defaults to \code{rounds = c(0,1)}.}

\item{max.iter}{number of iterations. Defaults to \code{max.iter = 10}.}

\item{na.rm}{Should missing variables be dropped? Defaults to \code{na.rm = FALSE}. See details for further information.}

\item{...}{future expansion.}
}
\value{
Objects of class "flowstat", which are tables with a "var" attribute giving the variance and a "statistic" attribute giving the type of flow.

These objects have methods for coef, vcov, SE, and cv.
}
\description{
Compute gross flows for data from complex surveys with repeated samples.
}
\details{
The \code{na.rm} option should be used cautiously. Usually, \code{NA} encoding has two possible meanings: a \emph{missing} information or
a \emph{not applicable} information. If the former should be dropped, while the latter should not. By default, if the table has an \code{NA}, it will
return an table with \code{NAs}. If \code{na.rm = TRUE}, the \code{NA} encoded responses are assumed to be missing at random, and Rojas et al. (2014)
method is applied.

It is important to distinguish missing responses from unnaplicable responses. This is feasible by  subsetting the design
for only applicable responses (with actual missing responses, if that is the case). For instance, suppose that we have two encoded variables:
(a) employed/unemployed; and (b) type of contract, with NAs if the response is missing or is unnaplicable. The answers of (a) define what are
the applicable responses for responses for (b). In this case, if we are going to analyze contract changes from people employed in the first round,
we should filter for people emplyed in both the first \emph{and} second rounds. This can be done using \code{subset}. Then, the remaining \code{NAs} are
actual missing responses.
}
\examples{

# load data
data( "artificial" )

# create surf design object
flowdes <-
  sfydesign( ids = ~0 ,
                 probs = ~ prob ,
                 data = list( dfa0 , dfa1 ) ,
                 nest = TRUE )

# gross flows
gross.flows <- svyflow( ~v0 , design = flowdes , flow.type = "gross" )
coef( gross.flows )
SE( gross.flows )

# net flows
net.flows <- svyflow( ~v0 , design = flowdes , flow.type = "net" )
coef( net.flows )
SE( net.flows )

}
\references{
ROJAS, H. A. G.; TRUJILLO, L.; SILVA, P. L. N. The estimation of gross flows in complex surveys with random nonresponse.
\emph{Survey Methodology}, v. 40, n. 2, p. 285–321, dec. 2014. URL \url{https://www150.statcan.gc.ca/n1/en/catalogue/12-001-X201400214113}.

LUMLEY, T. \emph{Complex Surveys:} A guide to analysis using R.
Hoboken: John Wiley & Sons, 2010. (Wiley Series in Survey Methodology). ISBN 978-0-470-28430-8.
}
\author{
Guilherme Jacob
}
