% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svyflow.R
\name{svyflow}
\alias{svyflow}
\alias{svyflow.survey.design2}
\alias{svyflow.svyrep.design}
\alias{svyflow.survflow.design}
\title{Summary statistics for repeated sample surveys}
\usage{
\method{svyflow}{survey.design2}(x, design, flow.type, rounds, max.iter, ...)

\method{svyflow}{svyrep.design}(x, design, flow.type, rounds, max.iter, ...)

\method{svyflow}{survflow.design}(x, design, ...)
}
\arguments{
\item{x}{A formula}

\item{design}{survflow.design object}

\item{flow.type}{type of flow to estimate: "gross" for counts and "net" for probabilities. Defaults to \code{flow.type = "gross"}.}

\item{rounds}{a vector of integers indicating which round to use. Defaults to \code{rounds = c(0,1)}.}

\item{max.iter}{number of iterations. Defaults to \code{max.iter = 10}.}

\item{...}{future expansion.}
}
\value{
Objects of class "flowstat", which are tables with a "var" attribute giving the variance and a "statistic" attribute giving the type of flow.

These objects have methods for coef , vcov, SE, and cv.
}
\description{
Compute gross flows for data from complex surveys with repeated samples.
}
\details{
...
}
\examples{

# load data
data( "initial" )
data( "final" )

# create survflow design object
flowdes <-
  svyflowdesign( ids = ~ upa ,
                 strata = ~ estrato ,
                 probs = ~ longprob ,
                 data.list = list( f.qtr , s.qtr ) ,
                 nest = TRUE )

# totals
svytotal( ~factor( vd4002 ) , flowdes , na.rm = TRUE ) # drops observations with missing in any rounds
svytotal( ~factor( vd4002 ) , flowdes , na.rm = TRUE , which.dataset = 0 ) # drops observations with missing in the first round
svytotal( ~factor( vd4002 ) , flowdes , na.rm = TRUE , which.dataset = 1 ) # drops observations with missing in the second round

# means
svymean( ~factor( vd4002 ) , flowdes , na.rm = TRUE ) # drops observations with missing in any rounds
svymean( ~factor( vd4002 ) , flowdes , na.rm = TRUE , which.dataset = 0 ) # drops observations with missing in the first round
svymean( ~factor( vd4002 ) , flowdes , na.rm = TRUE , which.dataset = 1 ) # drops observations with missing in the second round

}
\references{
Gutierrez, A., Trujillo, L. \& Silva, N. (2014). The estimation of gross flows in complex surveys with random nonresponse,
Survey Methodology 40(2), pp. 285-321.

Lumley, Thomas S. (2010). Complex Surveys: A Guide to Analysis Using R. Wiley Publishing.
}
\author{
Guilherme Jacob
}
