% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svyflow.R
\name{svyflow}
\alias{svyflow}
\alias{svyflow.survey.design2}
\alias{svyflow.svyrep.design}
\title{Gross flow estimation between categories}
\usage{
\method{svyflow}{survey.design2}(
  x,
  design,
  model = c("A", "B", "C", "D"),
  tol = 1e-04,
  maxit = 5000,
  verbose = FALSE,
  ...
)

\method{svyflow}{svyrep.design}(
  x,
  design,
  model = c("A", "B", "C", "D"),
  tol = 1e-04,
  maxit = 5000,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{a one-sided formula indicating a \emph{factor} variable.}

\item{design}{survey design object}

\item{model}{model for non-response. Possibilities: \code{"A", "B", "C", "D"}. Defaults to \code{model = "A"}.}

\item{tol}{Tolerance for iterative proportional fitting. Defaults to \code{1e-4}.}

\item{maxit}{Maximum number of iterations for iterative proportional fitting. Defaults to \code{maxit = 5000}.}

\item{verbose}{Print proportional fitting iterations. Defaults to \code{verbose = FALSE}.}

\item{...}{future expansion.}
}
\value{
Objects of class "flowstat", which are tables with a "var" attribute giving the variance and a "statistic" attribute giving the type of flow.

These objects have methods for coef, vcov, SE, and cv.
}
\description{
Compute gross flows for data from complex surveys with repeated samples.
}
\details{
It is important to distinguish "missing" responses from "unnaplicable" responses. This is feasible by subsetting the design
for only applicable responses (with actual missing responses, if that is the case). For instance, suppose that we have two variables encoded as
employed/unemployed, with NAs if the response is missing or is unnaplicable. An \code{NA} might be a person that did not respond \emph{or} a person
who was under the working-age at the time of the survey. It is important to distinguish across those, as only one of those cases is an
actual non-response. You could do that by looking for people who were in working age in any round, for instance. This can be done by using \code{subset},
as you should for a \code{survey design} object.
}
\examples{


# load library
library( survey )
library( surf )

# load data
data( "LFS79.0809" )

# create surf design object
lfs.des <- svydesign( ids = ~0 , probs = ~ prob , data = LFS79.0809 , nest = TRUE )

# flow estimates
estflows <- svyflow( ~y1+y2 , design = lfs.des )
coef( estflows$muij )
SE( estflows$muij )

}
\references{
GUTIERREZ, H. A.; TRUJILLO, L.; SILVA, P. L. N. The estimation of gross flows in complex surveys with random nonresponse.
\emph{Survey Methodology}, v. 40, n. 2, p. 285–321, dec. 2014. URL \url{https://www150.statcan.gc.ca/n1/en/catalogue/12-001-X201400214113}.

LUMLEY, T. \emph{Complex Surveys:} A guide to analysis using R.
Hoboken: John Wiley & Sons, 2010. (Wiley Series in Survey Methodology). ISBN 978-0-470-28430-8.
}
\author{
Guilherme Jacob
}
