% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svyflow.R
\name{svyflow}
\alias{svyflow}
\alias{svyflow.survey.design2}
\alias{svyflow.surflow.design}
\title{Gross flow estimation between categories}
\usage{
\method{svyflow}{survey.design2}(
  x,
  design,
  model = "A",
  rounds = c(0, 1),
  tol = 1e-06,
  verbose = FALSE,
  pij.zero = NULL
)

\method{svyflow}{surflow.design}(x, design, ...)
}
\arguments{
\item{x}{a formula indicating a \emph{factor} variable.}

\item{design}{surflow.design object}

\item{model}{model for non-response. Possibilities: \code{"A", "B", "C", "D"}. Defaults to \code{model = "A"}.}

\item{rounds}{a vector of integers indicating which round to use. Defaults to \code{rounds = c(0,1)}.}

\item{tol}{Tolerance for iterative proportional fitting. Defaults to \code{tol = 1e-6}.}

\item{verbose}{Print proportional fitting iterations. Defaults to \code{verbose = FALSE}.}

\item{pij.zero}{A logical matrix defining structural zeroes in transition matrix. Defaults to \code{pij.zero = NULL}.}

\item{...}{future expansion.}
}
\value{
Objects of class "flowstat", which are tables with a "var" attribute giving the variance and a "statistic" attribute giving the type of flow.

These objects have methods for coef, vcov, SE, and cv.
}
\description{
Compute gross flows for data from complex surveys with repeated samples.
}
\details{
The \code{na.rm} option should be used cautiously. Usually, \code{NA} encoding has two possible meanings: a \emph{missing} information or
a \emph{not applicable} information. If the former should be dropped, while the latter should not. By default, if the table has an \code{NA}, it will
return an table with \code{NAs}. If \code{na.rm = TRUE}, the \code{NA} encoded responses are assumed to be missing at random, and Rojas et al. (2014)
method is applied.

It is important to distinguish missing responses from unnaplicable responses. This is feasible by  subsetting the design
for only applicable responses (with actual missing responses, if that is the case). For instance, suppose that we have two encoded variables:
(a) employed/unemployed; and (b) type of contract, with NAs if the response is missing or is unnaplicable. The answers of (a) define what are
the applicable responses for responses for (b). In this case, if we are going to analyze contract changes from people employed in the first round,
we should filter for people emplyed in both the first \emph{and} second rounds. This can be done using \code{subset}. Then, the remaining \code{NAs} are
actual missing responses.
}
\examples{


# load library
library( surf )

# load data
data( "artificial" )

# create surf design object
flowdes <-
  sfydesign( ids = ~0 ,
                 probs = ~ prob ,
                 data = list( df0 , df1 ) ,
                 nest = TRUE )

# flow estimates
estflows <- svyflow( ~y , design = flowdes )
coef( estflows$muij )
SE( estflows$muij )

}
\references{
ROJAS, H. A. G.; TRUJILLO, L.; SILVA, P. L. N. The estimation of gross flows in complex surveys with random nonresponse.
\emph{Survey Methodology}, v. 40, n. 2, p. 285–321, dec. 2014. URL \url{https://www150.statcan.gc.ca/n1/en/catalogue/12-001-X201400214113}.

LUMLEY, T. \emph{Complex Surveys:} A guide to analysis using R.
Hoboken: John Wiley & Sons, 2010. (Wiley Series in Survey Methodology). ISBN 978-0-470-28430-8.
}
\author{
Guilherme Jacob
}
