% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svyflow.R
\name{svyflow}
\alias{svyflow}
\alias{svyflow.survey.design2}
\alias{svyflow.svyrep.design}
\alias{svyflow.surflow.design}
\title{Summary statistics for repeated sample surveys}
\usage{
\method{svyflow}{survey.design2}(x, design, flow.type, rounds, max.iter, ...)

\method{svyflow}{svyrep.design}(x, design, flow.type, rounds, max.iter, ...)

\method{svyflow}{surflow.design}(x, design, flow.type = NULL, rounds = c(0, 1), max.iter = 10, ...)
}
\arguments{
\item{x}{A formula}

\item{design}{surflow.design object}

\item{flow.type}{type of flow to estimate: "gross" for counts and "net" for probabilities. Defaults to \code{flow.type = "gross"}.}

\item{rounds}{a vector of integers indicating which round to use. Defaults to \code{rounds = c(0,1)}.}

\item{max.iter}{number of iterations. Defaults to \code{max.iter = 10}.}

\item{...}{future expansion.}
}
\value{
Objects of class "flowstat", which are tables with a "var" attribute giving the variance and a "statistic" attribute giving the type of flow.

These objects have methods for coef, vcov, SE, and cv.
}
\description{
Compute gross flows for data from complex surveys with repeated samples.
}
\details{
...
}
\examples{

# load data
data( "artificial" )

# create surf design object
flowdes <-
  sfydesign( ids = ~0 ,
                 probs = ~ prob ,
                 data.list = list( dfa0 , dfa1 ) ,
                 nest = TRUE )

# gross flows
gross.flows <- svyflow( ~v0 , design = flowdes , flow.type = "gross" )
coef( gross.flows )
SE( gross.flows )

# net flows
net.flows <- svyflow( ~v0 , design = flowdes , flow.type = "net" )
coef( net.flows )
SE( net.flows )

}
\references{
Gutierrez, A., Trujillo, L. \& Silva, N. (2014). The estimation of gross flows in complex surveys with random nonresponse,
Survey Methodology 40(2), pp. 285-321.

Lumley, Thomas S. (2010). Complex Surveys: A Guide to Analysis Using R. Wiley Publishing.
}
\author{
Guilherme Jacob
}
