% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform}
\alias{transform}
\title{This function performs a rigid body transformation of a surface file}
\usage{
transform(im, u = 0, v = 0, theta = 0, method = "imwarp")
}
\arguments{
\item{im}{\code{cimg} object}

\item{u, v}{numerical indicating the displacement in the x and y direction (in
pixels)}

\item{theta}{numerical indicating the angle of rotation (in degrees)}

\item{method}{method for interpolation (one of "taylor" or "imwarp")}
}
\value{
a transformed \code{cimg} object
}
\description{
This function performs a rigid body transformation of a surface file
}
\examples{
surf <- ground[[1]]
par(mfrow = c(1,2), mar = c(2,2,1,1))
plot(surf)
plot(transform(surf, theta = 45))
plot(transform(surf, 0.5, 0))
plot(transform(surf, 0.5, method = 'taylor'))
plot(transform(surf, theta = 1))
plot(transform(surf, theta = 1, method = 'taylor'))
}
