% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{tximport2}
\alias{tximport2}
\title{Import transcriptome quantification results}
\usage{
tximport2(
  files,
  type = c("none", "kallisto", "salmon", "sailfish", "rsem"),
  txIn = TRUE,
  txOut = FALSE,
  countsFromAbundance = c("no", "scaledTPM", "lengthScaledTPM"),
  tx2gene = NULL,
  reader = read.delim,
  geneIdCol,
  txIdCol,
  abundanceCol,
  countsCol,
  lengthCol,
  importer,
  collatedFiles,
  ignoreTxVersion = FALSE,
  quiet = F
)
}
\arguments{
\item{files}{character, files to transcriptome quantification results}
}
\value{
a list of length four, named as \code{abundance}, \code{counts}, \code{length}, and \code{countsFromAbundance}
}
\description{
Customized tximport from \code{DESeq2} and \code{RSEM} transcriptome quantification results.
}
